+++ oLink/App.config
@@ -1,11 +1,31 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-    <appSettings>
-		<add key="DB" value="server=.\SQLEXPRESS;Initial Catalog=oLink;User ID=sa;Password=pk94newuser" />
+	<configSections>
+		<section name="ResourceList" type="System.Configuration.NameValueFileSectionHandler,System, Version=1.0.3300.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+	</configSections>
+	<appSettings>
+		<add key="DB0" value="server=.\SQLEXPRESS;Initial Catalog=oLink;User ID=sas;Password=qNxzkWe3bxznfdf" />
+		<add key="S3_suffix" value="" />
+		<add key="localfiledir" value="File" />
+		<add key="localsitedir" value="Site" />
+		<add key="cfdn" value="cloudfront.net" />
+		<add key="cfdntesthtm" value="/Site/show.htm" />
+		<!--AWSProfileName is used to reference an account that has been registered with the SDK.
+If using AWS Toolkit for Visual Studio then this value is the same value shown in the AWS Explorer.
+It is also possible to register an account using the <solution-dir>/packages/AWSSDK-X.X.X.X/tools/account-management.ps1 PowerShell script
+that is bundled with the nuget package under the tools folder.
+
+		<add key="AWSProfileName" value="" />
+-->
 	</appSettings>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8" />
-    </startup>
+	<startup>
+		<supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
+	</startup>
+	<ResourceList>
+		<add key="jquery.min.js" value="https://cdn.jsdelivr.net/jquery/1.12.4/jquery.min.js" />
+		<add key="video-js.min.css" value="https://cdn.jsdelivr.net/npm/video.js@7.5.4/dist/video-js.min.css" />
+		<add key="video.min.js" value="https://cdn.jsdelivr.net/npm/video.js@7.5.4/dist/video.min.js" />
+	</ResourceList>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
