+++ oLink/FormLink.cs
@@ -1,34 +1,73 @@
-﻿using Ganss.XSS;
-using System;
-using System.Configuration;
-using System.Data;
-using System.Data.SqlClient;
-using System.Diagnostics;
-using System.Drawing;
+﻿using System;
+using System.Windows.Forms;
 using System.IO;
-using System.IO.Compression;
-using System.Net;
-using System.Reflection;
-using System.Security.Cryptography;
+using System.Drawing;
 using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading;
+using System.Net;
+using System.IO.Compression;
+using System.Data;
+using System.Data.SqlClient;
 using System.Web;
-using System.Windows.Forms;
+using System.Reflection;
+using System.Threading;
+using System.Diagnostics;
+using System.Collections.Generic;
+using System.Security.Cryptography;
+using System.Configuration;
+using Amazon.S3;
+using Amazon.S3.IO;
+using Amazon.S3.Model;
+using System.Linq;
+using Amazon.Runtime.CredentialManagement;
+using Amazon.Runtime;
+using Amazon.RDS;
+using Amazon;
+using Amazon.RDS.Model;
+using System.Collections.Specialized;
+using Amazon.CloudFront;
+using System.Net.Security;
+using System.Security.Cryptography.X509Certificates;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Amazon.S3.Transfer;
+using Ganss.XSS;
+using Amazon.CloudFront.Model;
+using System.Runtime.InteropServices;
 
 namespace oLink
 {
     public partial class FormLink : Form
     {
-        private static string sAppName = "oLink";
-        private static string sVersion = "001";
-        private static Point mp;
-        private static Random random = new Random();
-        private static string sUserAgent = "Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/65.0.3325.162 Safari/537.36";
-        public static string sWaited = "<div class=\"artn\"><p style=\"text-align:center;\">^wait^</p></div>";
-        public static string sNFound = "<div class=\"artn\"><p style=\"text-align:center;\">解析失败</p></div>";
-        public static string sConvert = "<div class=\"artn\"><p style=\"text-align:center;\">正在转换</p></div>";
+        static private string sAppName = "oLink";
+        static private string sVersion = "001";
+        static private Point mp;
+        static private Random random = new Random();
+        static private string sUserAgent = "Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/65.0.3325.162 Safari/537.36";
+        static public string sWaited = "<div class=\"artn\"><p style=\"text-align:center;\">^wait^</p></div>";
+        static public string sNFound = "<div class=\"artn\"><p style=\"text-align:center;\">解析失败</p></div>";
+        static public string sConvert = "<div class=\"artn\"><p style=\"text-align:center;\">正在转换</p></div>";
 
+        private static readonly byte[] BMP = { 66, 77 };
+        private static readonly byte[] DOC = { 208, 207, 17, 224, 161, 177, 26, 225 };
+        private static readonly byte[] EXE_DLL = { 77, 90 };
+        private static readonly byte[] GIF = { 71, 73, 70, 56 };
+        private static readonly byte[] ICO = { 0, 0, 1, 0 };
+        private static readonly byte[] JPG = { 255, 216, 255 };
+        private static readonly byte[] MP3 = { 255, 251, 48 };
+        private static readonly byte[] OGG = { 79, 103, 103, 83, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0 };
+        private static readonly byte[] PDF = { 37, 80, 68, 70, 45, 49, 46 };
+        private static readonly byte[] PNG = { 137, 80, 78, 71, 13, 10, 26, 10, 0, 0, 0, 13, 73, 72, 68, 82 };
+        private static readonly byte[] RAR = { 82, 97, 114, 33, 26, 7, 0 };
+        private static readonly byte[] SWF = { 70, 87, 83 };
+        private static readonly byte[] TIFF = { 73, 73, 42, 0 };
+        private static readonly byte[] TORRENT = { 100, 56, 58, 97, 110, 110, 111, 117, 110, 99, 101 };
+        private static readonly byte[] TTF = { 0, 1, 0, 0, 0 };
+        private static readonly byte[] WAV_AVI = { 82, 73, 70, 70 };
+        private static readonly byte[] WMV_WMA = { 48, 38, 178, 117, 142, 102, 207, 17, 166, 217, 0, 170, 0, 98, 206, 108 };
+        private static readonly byte[] ZIP_DOCX = { 80, 75, 3, 4 };
+
+        private static readonly string encrypt_fileext = ".dat";
         public FormLink()
         {
             InitializeComponent();
@@ -36,9 +75,10 @@ namespace oLink
 
         private void FormMain_Load(object sender, EventArgs e)
         {
+            
             try
             {
-                ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
+                //ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
                 //ServicePointManager.ServerCertificateValidationCallback = delegate { return true; };
 
                 DataGridViewColumn cc = new DataGridViewCheckBoxColumn();
@@ -70,12 +110,33 @@ namespace oLink
                 dataGridView1.AllowUserToResizeRows = false;
                 foreach (DataGridViewColumn col in dataGridView1.Columns) col.SortMode = DataGridViewColumnSortMode.NotSortable;
 
-                ShowConfig();
-                ShowDataGridView1();
+                //ShowConfig();
+                //ShowDataGridView1();
+
+                listprofilenames();
+                PrepareResourceFiles4S3();
+
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
+            finally
+            {
+                ShowhideDataTabs(false);
+            }
         }
 
+        private void PrepareResourceFiles4S3()
+        {
+            OrganizeLocalFolders();
+            string localfiledirStr = ConfigurationManager.AppSettings["localfiledir"];
+
+            NameValueCollection resrc_section = (NameValueCollection)ConfigurationManager.GetSection("ResourceList");
+
+            foreach (string keystring in resrc_section.Keys)
+            {
+                string url_value = resrc_section.GetValues(keystring)[0];
+                DownloadHtml(url_value, "", "", localfiledirStr + @"\" + keystring);
+            }
+        }
         private void buttonClose_Click(object sender, EventArgs e)
         {
             Close();
@@ -141,29 +202,40 @@ namespace oLink
         }
 
         private void button2_Click(object sender, EventArgs e)
-        {
+        {//save button on tabpage1, below Note
             try
             {
-                string s1 = ooData.sG10配置Set
-                    .Replace("^S3Id^", GetSqlParam(textBoxS3Id.Text.Trim()))
-                    .Replace("^S3Key^", GetSqlParam(textBoxS3Key.Text.Trim()))
-                    .Replace("^S3Bucket^", GetSqlParam(textBoxS3Bucket.Text.Trim()))
-                    .Replace("^Name^", GetSqlParam(textBoxName.Text.Trim()))
-                    .Replace("^Note^", GetSqlParam(textBoxNote.Text.Trim()));
-                ExecuteSQL(s1);
+                string s1 = ooData.sG10配置Set_p;
+                List<SqlParameter> sqlparms = new List<SqlParameter>()
+                {
+                    new SqlParameter() {ParameterName = "@S3Id", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(textBoxS3Id.Text.Trim())},
+                    new SqlParameter() {ParameterName = "@S3Key", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(textBoxS3Key.Text.Trim())},
+                    new SqlParameter() {ParameterName = "@S3Bucket", SqlDbType = SqlDbType.NChar, Value =  GetSqlParam(textBoxS3Bucket.Text.Trim())},
+                    new SqlParameter() {ParameterName = "@Name", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(textBoxName.Text.Trim())},
+                    new SqlParameter() {ParameterName = "@Note", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(textBoxNote.Text.Trim())}
+                };
+                ExecuteSQL(s1, null, sqlparms);
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
+            //"Add" "Website Url" button
             try
             {
                 string s1 = textBox3.Text.Trim();
                 if (s1 == "" || (!s1.StartsWith("https://") && !s1.StartsWith("http://"))) return;
-                string s4 = ooData.sG10事物Add
-                    .Replace("^名称^", GetSqlParam(s1));
-                ExecuteSQL(s4);
+                //string s4 = ooData.sG10事物Add
+                //    .Replace("^名称^", GetSqlParam(s1));
+
+                string s4 = ooData.sG10事物Add_p;
+                List<SqlParameter> sqlparms = new List<SqlParameter>()
+                {
+                    new SqlParameter() {ParameterName = "@名称", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(s1)}
+                };
+                //ExecuteSQL(s4);
+                ExecuteSQL(s4, null, sqlparms);
                 ShowDataGridView1();
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
@@ -176,17 +248,25 @@ namespace oLink
                 for (int i = 0; i < dataGridView1.Rows.Count; i++)
                 {
                     if (Convert.ToBoolean(dataGridView1.Rows[i].Cells[0].Value) == false) continue;
-                    string s4 = ooData.sG10事物Del
-                        .Replace("^序号^", GetSqlParam(dataGridView1.Rows[i].Cells[1].Value.ToString()));
-                    ExecuteSQL(s4);
+                    //string s4 = ooData.sG10事物Del
+                    //    .Replace("^序号^", GetSqlParam(dataGridView1.Rows[i].Cells[1].Value.ToString()));
+
+                    string s4 = ooData.sG10事物Del_p;
+                    List<SqlParameter> sqlparms = new List<SqlParameter>()
+                    {
+                        new SqlParameter() {ParameterName = "@序号", SqlDbType = SqlDbType.Int, Value = GetSqlParam(dataGridView1.Rows[i].Cells[1].Value.ToString())}
+                    };
+                    ExecuteSQL(s4, null, sqlparms);
                 }
                 ShowDataGridView1();
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
+
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
+            //Retrieve button
             Thread thread = new Thread(Retrieve);
             thread.IsBackground = true;
             thread.Start();
@@ -196,6 +276,7 @@ namespace oLink
         {
             try
             {
+                ShowMsgD("Retrieve Start");
                 DataSet ds = new DataSet();
                 ExecuteSQL(ooData.sG10事物All, ds);
                 for (int i = 0; i < ds.Tables[0].Rows.Count; i++)
@@ -208,12 +289,20 @@ namespace oLink
                     if (s标题 != "")
                     {
                         if (s封面 != "") s封面 = GetFile(s封面);
-                        string sql = ooData.sG10事物Set
-                            .Replace("^标题^", GetSqlParam(s标题))
-                            .Replace("^摘要^", GetSqlParam(s摘要))
-                            .Replace("^封面^", GetSqlParam(s封面))
-                            .Replace("^名称^", GetSqlParam(s2));
-                        string s3 = ExecuteSQLResult(sql);
+                        //string sql = ooData.sG10事物Set
+                        //    .Replace("^标题^", GetSqlParam(s标题))
+                        //    .Replace("^摘要^", GetSqlParam(s摘要))
+                        //    .Replace("^封面^", GetSqlParam(s封面))
+                        //    .Replace("^名称^", GetSqlParam(s2));
+                        string sql = ooData.sG10事物Set_p;
+                        List<SqlParameter> sqlparms = new List<SqlParameter>()
+                        {
+                            new SqlParameter() {ParameterName = "@标题", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(s标题)},
+                            new SqlParameter() {ParameterName = "@摘要", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(s摘要)},
+                            new SqlParameter() {ParameterName = "@封面", SqlDbType = SqlDbType.NVarChar, Value =  GetSqlParam(s封面)},
+                            new SqlParameter() {ParameterName = "@名称", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(s2)}
+                        };
+                        string s3 = ExecuteSQLResult(sql, sqlparms);
 
                         string s4 = GetHtmlPlayer(s2, s封面);
                         s4 = s4.Replace("[链接[", "<a href=\"javascript:void(0);\" onclick=\"javascript:Load(encodeURI('")
@@ -226,7 +315,7 @@ namespace oLink
                 ds = null;
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
-            ShowMsgD("Done");
+            ShowMsgD("Retrieve Done");
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -240,6 +329,7 @@ namespace oLink
         {
             try
             {
+                ShowMsgD("Reconstruct Start");
                 string s6 = "";
                 DataSet ds = new DataSet();
                 ExecuteSQL(ooData.sG10事物All, ds);
@@ -266,7 +356,7 @@ namespace oLink
                 SaveFile(s6, "Site/olHome.htm");
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
-            ShowMsgD("Done");
+            ShowMsgD("Reconstruct Done");
         }
 
         private void GetArticle(string name0, out string sTitle, out string sAdstract, out string sCover, out string sTime)
@@ -308,9 +398,17 @@ namespace oLink
         {
             try
             {
-                string s1 = ooData.sG10文件Add
-                    .Replace("^名称^", GetSqlParam(name));
-                string s2 = ExecuteSQLResult(s1);
+                //string s1 = ooData.sG10文件Add
+                //    .Replace("^名称^", GetSqlParam(name));
+                string s1 = ooData.sG10文件Add_p;
+                List<SqlParameter> sqlparms = new List<SqlParameter>()
+                {
+                    new SqlParameter() {ParameterName = "@名称", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(name)}
+                };
+
+                //string s2=ExecuteSQLResult(s1);
+                string s2 = ExecuteSQLResult(s1, sqlparms);
+
                 string s4 = "File/" + int.Parse(s2).ToString("D6");
                 if (name.EndsWith(".jpg")) s4 += ".jpg";
                 if (name.EndsWith(".jpeg")) s4 += ".jpg";
@@ -337,34 +435,74 @@ namespace oLink
 
         private void Upload()
         {
-            ShowMsgD("Start");
+            ShowMsgD("Start Upload");
             try
             {
+                ClearDangerousFilesinLocalFolders();    //File check before upload 8/2022
                 string s0 = Path.GetDirectoryName(Application.ExecutablePath);
-                string s = @"set AWS_ACCESS_KEY_ID=^S3Id^
-set AWS_SECRET_ACCESS_KEY=^S3Key^
-aws s3 sync C:\oLink\Site s3://^S3Bucket^/Site --acl public-read --region eu-west-1
-aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-west-1"
-.Replace("^S3Id^", textBoxS3Id.Text.Trim())
-.Replace("^S3Key^", textBoxS3Key.Text.Trim())
-.Replace("^S3Bucket^", textBoxS3Bucket.Text.Trim());
-                Process p = new Process();
-                p.StartInfo.FileName = "cmd.exe";
-                p.StartInfo.UseShellExecute = false;
-                p.StartInfo.RedirectStandardInput = true;
-                p.StartInfo.RedirectStandardOutput = true;
-                p.StartInfo.RedirectStandardError = true;
-                p.StartInfo.CreateNoWindow = true;
-                p.OutputDataReceived += new DataReceivedEventHandler(OnOutputDataReceived);
-                p.Start();
-                p.BeginOutputReadLine();
-                p.StandardInput.WriteLine(s);
-                p.StandardInput.WriteLine("exit");
-                p.WaitForExit();
-                if (p.ExitCode != 0) ShowMsgD(p.StandardError.ReadToEnd());
+                AmazonS3Client s3cli = AWSConnections.getAWSS3Cli(tbxAccKey.Text, tbxSecKey.Text);
+                string st_myS3name = AWSConnections.getAWSS3Client(tbxAccKey.Text, tbxSecKey.Text);
+
+                if (s3cli == null)
+                {
+                    string s = @"set AWS_ACCESS_KEY_ID=^S3Id^
+                                set AWS_SECRET_ACCESS_KEY=^S3Key^
+                                aws s3 sync C:\oLink\Site s3://^S3Bucket^/Site --acl public-read --region eu-west-1
+                                aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-west-1"
+                                .Replace("^S3Id^", textBoxS3Id.Text.Trim())
+                                .Replace("^S3Key^", textBoxS3Key.Text.Trim())
+                                .Replace("^S3Bucket^", textBoxS3Bucket.Text.Trim());
+                    Process p = new Process();
+                    p.StartInfo.FileName = "cmd.exe";
+                    p.StartInfo.UseShellExecute = false;
+                    p.StartInfo.RedirectStandardInput = true;
+                    p.StartInfo.RedirectStandardOutput = true;
+                    p.StartInfo.RedirectStandardError = true;
+                    p.StartInfo.CreateNoWindow = true;
+                    p.OutputDataReceived += new DataReceivedEventHandler(OnOutputDataReceived);
+                    p.Start();
+                    p.BeginOutputReadLine();
+                    p.StandardInput.WriteLine(s);
+                    p.StandardInput.WriteLine("exit");
+                    p.WaitForExit();
+                    if (p.ExitCode != 0) ShowMsgD(p.StandardError.ReadToEnd());
+                }
+                else
+                {
+                    string remote_folderPath = "Site";
+                    string Local_folderpath = s0 + @"\Site\";// @"C:\oLink\Site";
+                    //S3DirectoryInfo directory = new S3DirectoryInfo(s3cli, st_myS3name, remote_folderPath);
+                    //directory.MoveFromLocal(Local_folderpath);
+                    var transferUtility = new TransferUtility(s3cli);
+                    TransferUtilityUploadDirectoryRequest tuudr_request = new TransferUtilityUploadDirectoryRequest()
+                    {
+                        BucketName = st_myS3name,
+                        Directory = Local_folderpath,
+                        KeyPrefix = remote_folderPath
+                    };
+                    transferUtility.UploadDirectory(tuudr_request);
+
+
+                    remote_folderPath = "File";
+                    Local_folderpath = s0 + @"\File\";// @"C:\oLink\File";
+                    //directory = new S3DirectoryInfo(s3cli, st_myS3name, remote_folderPath);
+                    //directory.MoveFromLocal(Local_folderpath);
+
+                    tuudr_request = new TransferUtilityUploadDirectoryRequest()
+                    {
+                        BucketName = st_myS3name,
+                        Directory = Local_folderpath,
+                        KeyPrefix = remote_folderPath
+                    };
+                    transferUtility.UploadDirectory(tuudr_request);
+
+                    transferUtility.Dispose();
+                }
+
+
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
-            ShowMsgD("Done");
+            ShowMsgD("Done Upload");
         }
 
         private void OnOutputDataReceived(object Sender, DataReceivedEventArgs e)
@@ -391,17 +529,84 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
         {
             try
             {
-                string url = "https://s3.amazonaws.com/^S3Bucket^/Site/show.htm?ag=olHome&pin=^random^#olHome"
-                    .Replace("^S3Bucket^", textBoxS3Bucket.Text.Trim())
+                string cf_url = retrievecfurl();
+                string url = "";
+                string cfsitedn = ConfigurationManager.AppSettings["cfdn"];
+                if (!cf_url.ToUpper().Contains(cfsitedn.ToUpper()) || !ckbcloudfront.Checked)
+                {
+                    url = "https://s3.amazonaws.com/^S3Bucket^/Site/show.htm?ag=olHome&pin=^random^#olHome"
+                      .Replace("^S3Bucket^", textBoxS3Bucket.Text.Trim())
+                      .Replace("^random^", GetRandom());
+                }
+                else
+                {
+                    url = "https://" + cf_url + "/Site/show.htm?ag=olHome&pin=^random^#olHome"
                     .Replace("^random^", GetRandom());
-                string s = GetHtmlMethod("GET", "https://is.gd/create.php?format=simple&url=" + EnUrlSymbol(url), "", "", "", "");
+                }
+                TogglePublicAccessBlock(ckbcloudfront.Checked);
+
+                string s = url;
+                if (ckbShortURL.Checked)
+                    s = GetHtmlMethod("GET", "https://is.gd/create.php?format=simple&url=" + EnUrlSymbol(url), "", "", "", "");
                 ShowMsgD(s);
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
         }
 
+        private string retrievecfurl()
+        {
+            string awsAccessKey = tbxAccKey.Text;
+            string awsSecretAccessKey = tbxSecKey.Text;
+            try
+            {
+                AmazonCloudFrontClient acfcli = new AmazonCloudFrontClient(awsAccessKey, awsSecretAccessKey);
+
+                if (acfcli != null)
+                {
+                    var dl = acfcli.ListDistributions();
+                    if (dl.ContentLength > 0)
+                    {
+                        foreach (var ditem in dl.DistributionList.Items)
+                        {
+                            foreach (var dorg in ditem.Origins.Items)
+                            {
+                                if (dorg.DomainName.ToUpper().Contains(tbxS3buckets.Text.ToUpper()))
+                                {
+                                    //if (testurlaccessible(ditem.DomainName))
+                                    return ditem.DomainName;
+                                }
+                            }
+                        }
+                    }
+                    return "";
+                }
+            }
+            catch (Exception ex)
+            {
+                Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message);
+            }
+            return "";
+        }
+
+        private bool testurlaccessible(string url)
+        {
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Ssl3;
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls12;// | SecurityProtocolType.Tls13;
+            //ServicePointManager.ServerCertificateValidationCallback += new RemoteCertificateValidationCallback(AlwaysGoodCertificate);
+            ServicePointManager.ServerCertificateValidationCallback += (sender, certificate, chain, errors) => { return CertificateValidationCallBack(sender, certificate, chain, errors); };
+            HttpClient client = new HttpClient();
+
+            HttpResponseMessage checkingResponse = client.GetAsync(url).Result;
+            if (!checkingResponse.IsSuccessStatusCode)
+            {
+                return false;
+            }
+            return true;
+        }
+
+
         // Player
-        public static string GetVideoPlayer(string url, string track = "", string cover = "", string myip = "", bool bDownload = true)
+        static public string GetVideoPlayer(string url, string track = "", string cover = "", string myip = "", bool bDownload = true)
         {
             if (url == "") return "";
             if (url.StartsWith("https://player.vimeo.com/video/")) url = url.Replace("https://player.vimeo.com/", "https://vimeo.com/");
@@ -413,125 +618,12 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             string sTrack = "";
             string sTrackZh = "", sTrackEn = "";
             string sTrackZhSrc = "", sTrackEnSrc = "";
-            //if (track != "" && track.EndsWith(".vtt"))
-            //{
-            //    sTrackZh = "zh";
-            //    sTrackZhSrc = EnUrl2(track);
-            //    sTrack += "\r\n<track kind=\"subtitles\" label=\"中文\" srclang=\"zh\" src=\"" + sTrackZhSrc + "\" default>";
-            //}
-
-            //if (url.StartsWith("https://www.youtube.com/") || url.StartsWith("https://youtu.be/") || url.StartsWith("https://www.youtube.com/embed/"))
-            //{
-            //    url = url.Replace("https://www.youtube.com/watch?v=", "").Replace("https://youtu.be/", "").Replace("https://www.youtube.com/embed/", "");
-            //    if (url.Length > 11) url = url.Substring(0, 11);
-
-            //    string sGetYtb = "";
-            //    string[] ssGetYtb = sGetYtb.Split(',');
-            //    string uinfo = "https://www.youtube.com/get_video_info?eurl=&sts=^sSts^&video_id=".Replace("^sSts^", ssGetYtb[0]) + url;
-            //    string s1 = GetHtml(uinfo, true);
-            //    if (s1.Contains("%22status%22%3A%22UNPLAYABLE%22")) s1 = GetHtml(uinfo + "&el=detailpage", true); // "&el=embedded", "&el=vevo", ""
-            //    if (s1 == "") return sNFound;
-
-            //    // Wait
-            //    Match mWait = new Regex("(?<=%22status%22%3A%22LIVE_STREAM_OFFLINE%22%2C%22reason%22%3A%22)([\\S\\s]*?)(?=%22)").Match(s1);
-            //    if (mWait.Success) return sWaited.Replace("^wait^", HttpUtility.UrlDecode(mWait.Value)
-            //        + "<img class='artl' src='" + EnUrl2("https://i.ytimg.com/vi/" + url + "/maxresdefault.jpg") + "'>");
-
-            //    // m3u8
-            //    Match mM3u8 = new Regex("(?<=%22hlsManifestUrl%22%3A%22)([\\S\\s]*?)(?=%22)").Match(s1);
-            //    if (mM3u8.Success)
-            //    {
-            //        string sM3u8 = GetHtml(mM3u8.Value.Replace("%3A", ":").Replace("%2F", "/").Replace("%252C", ",").Replace("%253D", "="));
-            //        Match mM3u8_360 = new Regex("(?<=RESOLUTION=640x[\\S\\s]*?)(https[\\S]*?)(?=\\s)").Match(sM3u8);
-            //        if (!mM3u8_360.Success) return sNFound;
-            //        string m3u8 = GetHtml(mM3u8_360.Value);
-            //        string[] m3u8s = m3u8.Split('\n'); //Log(url + " " + m3u8s.Length+" "+ mM3u8_360.Value);
-            //        if (m3u8s.Length > 100 || m3u8 == "") return sConvert;
-            //        sM3u8 = "http://or9a.odisk.org/oo.aspx?name=get_m3u8&ag=" + HttpUtility.UrlEncode(mM3u8_360.Value)
-            //            + "&myip=" + myip + "&type=play.m3u8";
-            //        return GetM3u8Player(sM3u8);
-            //    }
-
-            //    Match mTrackList = new Regex("(?<=%22captionTracks%22%3A%5B)([\\S\\s]*?)(?=%5D)").Match(s1);
-            //    string sTrackList = mTrackList.Value
-            //        .Replace("%3A", ":").Replace("%2F", "/").Replace("%3F", "?").Replace("%26", "&")
-            //        .Replace("%3D", "=").Replace("%2C", ",").Replace("%2B", "+")
-            //        .Replace("%7B", "{").Replace("%7D", "}").Replace("%22", "\"")
-            //        .Replace("%252C", ",").Replace("%5C", "\\").Replace("\\u0026", "&");
-            //    MatchCollection mc = new Regex("\"name\":\\{([\\S\\s]*?)\\}").Matches(sTrackList);
-            //    foreach (Match m in mc)
-            //        sTrackList = sTrackList.Replace(m.Value, m.Value.Replace("\"name\":{", "").Replace("}", ""));
-            //    if
-            //            (sTrackList.Contains("\"vssId\":\".zh-CN\"")) sTrackZh = ".zh-CN"; // lang_code
-            //    else if (sTrackList.Contains("\"vssId\":\".zh\"")) sTrackZh = ".zh";
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-Hans\"")) sTrackZh = ".zh-Hans";
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-HK\"")) sTrackZh = ".zh-HK";
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-TW\"")) sTrackZh = ".zh-TW";
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-SG\"")) sTrackZh = ".zh-SG"; // 新加坡
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-Hant\"")) sTrackZh = ".zh-Hant";
-            //    else if (sTrackList.Contains("\"vssId\":\".zh-Hans.GZolFkXAZLc\"")) sTrackZh = ".zh-Hans.GZolFkXAZLc";
-            //    else if (sTrackList.Contains("\"vssId\":\".nan-TW\"")) sTrackZh = ".nan-TW"; // Min Nan Chinese (Taiwan)
-            //    if
-            //            (sTrackList.Contains("\"vssId\":\".en\"")) sTrackEn = ".en";
-            //    else if (sTrackList.Contains("\"vssId\":\".en-GB\"")) sTrackEn = ".en-GB";
-            //    else if (sTrackList.Contains("\"vssId\":\".en-US\"")) sTrackEn = ".en-US";
-            //    else if (sTrackList.Contains("\"vssId\":\"a.en\"")) sTrackEn = "a.en";
-            //    else if (sTrackList.Contains("\"vssId\":\"a.ja\"")) sTrackEn = "a.ja";
-            //    else if (sTrackList.Contains("\"vssId\":\"a.de\"")) sTrackEn = "a.de";
-            //    else if (sTrackList.Contains("\"vssId\":\"a.ru\"")) sTrackEn = "a.ru";
-            //    else if (sTrackList.Contains("\"vssId\":\"a.fr\"")) sTrackEn = "a.fr"; // 法
-            //    mc = new Regex("\\{([\\S\\s]*?)\\}").Matches(sTrackList);
-            //    if (sTrackZh != "" && sTrackZhSrc == "")
-            //        foreach (Match m in mc)
-            //            if (m.Value.Contains("\"vssId\":\"" + sTrackZh + "\""))
-            //            {
-            //                Match m2 = new Regex("(?<=\"baseUrl\":\")([\\S\\s]*?)(?=\")").Match(m.Value);
-            //                sTrackZhSrc = EnUrl2(m2.Value + "&fmt=vtt&type=.vtt");
-            //                sTrack += "\r\n<track kind=\"subtitles\" label=\"^origin^\" srclang=\"^code^\" src=\"^src^\" default>"
-            //                .Replace("^origin^", "原始中文").Replace("^code^", sTrackZh).Replace("^src^", sTrackZhSrc);
-            //            }
-            //    if (sTrackEn != "")
-            //        foreach (Match m in mc)
-            //            if (m.Value.Contains("\"vssId\":\"" + sTrackEn + "\""))
-            //            {
-            //                Match m2 = new Regex("(?<=\"baseUrl\":\")([\\S\\s]*?)(?=\")").Match(m.Value);
-            //                if (sTrackZh == "")
-            //                {
-            //                    sTrackZh = "zh-Hans";
-            //                    sTrackZhSrc = EnUrl2(m2.Value + "&fmt=vtt&tlang=zh-Hans&type=.vtt");
-            //                    sTrack += "\r\n<track kind=\"subtitles\" label=\"^origin^\" srclang=\"zh\" src=\"^src^\" default>"
-            //                    .Replace("^origin^", "自动中文").Replace("^src^", sTrackZhSrc);
-            //                }
-            //                sTrackEnSrc = EnUrl2(m2.Value + "&fmt=vtt&type=.vtt");
-            //                sTrack += "\r\n<track kind=\"subtitles\" label=\"^origin^\" srclang=\"en\" src=\"^src^\">"
-            //                .Replace("^origin^", (sTrackEn == "a.en" ? "自动外文" : "原始外文")).Replace("^src^", sTrackEnSrc);
-            //            }
-
-            //    // video & audio
-            //    Match m1 = new Regex("(?<=%22formats%22%3A%5B)([\\S]*?)(?=%5D)").Match(s1);
-            //    if (!m1.Success) return sNFound;
-            //    string s2 = m1.Value
-            //        .Replace("%3A", ":").Replace("%2F", "/").Replace("%3F", "?").Replace("%26", "&")
-            //        .Replace("%3D", "=").Replace("%2C", ",").Replace("%2B", "+")
-            //        .Replace("%7B", "{").Replace("%7D", "}").Replace("%22", "\"");
-            //    string[] ss1 = s2.Split(new string[] { "},{" }, StringSplitOptions.RemoveEmptyEntries);
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":22")) { sVideoV = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":18")) { sVideoM = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":45")) { sVideoV2 = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":43")) { sVideoM2 = GetYoutube(s); break; }
-            //    m1 = new Regex("(?<=%22adaptiveFormats%22%3A%5B)([\\S]*?)(?=%5D)").Match(s1);
-            //    s2 = m1.Value
-            //        .Replace("%3A", ":").Replace("%2F", "/").Replace("%3F", "?").Replace("%26", "&")
-            //        .Replace("%3D", "=").Replace("%2C", ",").Replace("%2B", "+")
-            //        .Replace("%7B", "{").Replace("%7D", "}").Replace("%22", "\"");
-            //    ss1 = s2.Split(new string[] { "},{" }, StringSplitOptions.RemoveEmptyEntries);
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":140")) { sAudio140 = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":171")) { sAudio171 = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":249")) { sAudio249 = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":250")) { sAudio250 = GetYoutube(s); break; }
-            //    foreach (string s in ss1) if (s.Contains("\"itag\":251")) { sAudio251 = GetYoutube(s); break; }
-            //    url = EnUrl2(sVideoM);
-            //}
+            if (track != "" && track.EndsWith(".vtt"))
+            {
+                sTrackZh = "zh";
+                sTrackZhSrc = EnUrl2(track);
+                sTrack += "\r\n<track kind=\"subtitles\" label=\"中文\" srclang=\"zh\" src=\"" + sTrackZhSrc + "\" default>";
+            }
 
             url = GetEnUrl(url);
             string id = GetRandomNum();
@@ -651,7 +743,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
         src=""^url^"" ^poster^>
     </video>
     <div class=""navv""><div class=""nav"" style=""width:100%; height:28px; position:absolute; top:0;"">
-        <ul class=""nav__menu"" style=""text-align:right;"">
+        <ul class=""nav__menu"" style=""text-align:right;"">           
             <li class=""nav__menu-item""><a>速度</a>
                 <ul class=""nav__submenu"">
                     <li class=""nav__submenu-item"" onclick='javascript:document.getElementById(""v^id^"").playbackRate=0.5;'><a>0.5</a></li>
@@ -680,41 +772,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             .Replace("^poster^", (cover != "" ? "poster=\"" + cover + "\"" : ""));
         }
 
-    //    private static string GetSoundPlayer(string url, string cover = "")
-    //    {
-    //        url = EnUrl2(url);
-    //        return @"
-    //<video controls ^autoplay^ preload=""auto"" crossorigin=""anonymous"" style=""width:100%; height:180px; background:#faf2e1;""
-    //    src=""^src^"" ^poster^>
-    //</video>"
-    //        .Replace("^autoplay^", (cover != "" ? "" : "")).Replace("^src^", GetHtmlParam(url)) /// autoplay
-    //        .Replace("^poster^", (cover != "" ? "poster='" + GetHtmlParam(cover) + "'" : ""));
-    //    }
-
-        private static string GetM3u8Player(string url)
-        {
-            url = GetEnUrl(url);
-            return @"
-    <video id=""p@id"" class=""video-js"" controls autoplay width=""100%"" data-setup=""{}"">
-        <source src=""@src"" type=""application/x-mpegURL"">
-    </video>".Replace("@id", GetRandomNum()).Replace("@src", url); /// no poster= poster=""""
-        }
-
-//        private static string GetFlashPlayer(string url)
-//        {
-//            url = EnUrl2(url);
-//            return @"
-//<link rel=""stylesheet"" href=""/src/functional.css"">
-//<style> .flowplayer {background-color:#ffffd0;} </style>
-//<script src=""/src/flowplayer.min.js""></script>
-//<div class=""flowplayer"">
-//    <video>
-//        <source type=""video/flash"" src=""@src"">
-//    </video>
-//</div>".Replace("@src", GetHtmlParam(url));
-//        }
-
-        private static string GetImagePlayerTop(string url)
+        static private string GetImagePlayerTop(string url)
         {
             url = GetEnUrl(url);
             return @"<div class='videocontainer'>
@@ -728,49 +786,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             .Replace("^url^", url);
         }
 
-//        private static string GetImagePlayer(string url)
-//        {
-//            url = EnUrl2(url);
-//            return @"<div class='videocontainer' style='max-width:580px; margin:6px auto;'>
-//    <img style='width:100%;' src='^url^'/>
-//    <div class=""navv""><div class=""nav"" style=""width:100%; height:28px; position:absolute; top:0;"">
-//        <ul class=""nav__menu"" style=""text-align:right;"">
-//            <li class=""nav__menu-item""><a href=""javascript:void(0);"" onclick=""javascript:Save('^url^');"">下载</a></li>
-//        </ul>
-//    </div></div>
-//</div>"
-//            .Replace("^url^", GetHtmlParam(url));
-//        }
-
-        //private static string GetDownloadPlayer(string url)
-        //{
-        //    url = EnUrl2(url);
-        //    return "\r\n  <div class='buttoncontainer'><a href='javascript:void(0);' onclick=\"javascript:Save('"
-        //        + GetHtmlParam(url) + "');\"><div class='button'>点击下载</div></a></div><div class='lisw'></div>";
-        //}
-
-        //private static string GetTwitter(string co, string cover, string title)
-        //{
-        //    string r = "";
-        //    try
-        //    {
-        //        string[] ss = co.Split(new string[] { "[分隔[时间]分隔]" }, StringSplitOptions.RemoveEmptyEntries);
-        //        foreach (string s in ss)
-        //        {
-        //            string[] ss2 = s.Split(new string[] { "[分隔[推文]分隔]" }, StringSplitOptions.None);
-        //            r +=
-        //              "\r\n    <div class='lisu'>"
-        //            + "\r\n      <div class='picv'><img class='imgh' src='" + cover + "'></div>"
-        //            + "\r\n      <div class='artv'><p class='namv'>" + "<font class='count'>" + title + " " + ss2[0] + "</font></p>"
-        //            + "\r\n      </div>"
-        //            + "\r\n    </div>";
-        //        }
-        //    }
-        //    catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name, ex.Message); }
-        //    return r;
-        //}
-
-        private static string GetEnUrl(string url)
+        static private string GetEnUrl(string url)
         {
             return EnUrl2(url);
         }
@@ -1049,107 +1065,105 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
 
         public string HtmlDel2(string content, string url)
         {
-            // filter example:
-            //content += @"<><<>img src=x onerror=""alert('img tag/script bypassed filtering'); "">";
             var sanitizer = new HtmlSanitizer();
             try
             {
                 content = sanitizer.Sanitize(content, url);
+                content = HtmlDel(content, "(<head)([\\S\\s]*?)(</head>)");
+                content = HtmlDel(content, "(<script)([\\S\\s]*?)(</script>)");
+                content = HtmlDel(content, "(<noscript)([\\S\\s]*?)(</noscript>)");
+                content = HtmlDel(content, "(<style)([\\S\\s]*?)(</style>)");
+                content = HtmlDel(content, "(<ins)([\\S\\s]*?)(</ins>)");
+
+                MatchCollection mc1 = new Regex("(<a )([\\S\\s]*?)(</a>)").Matches(content);
+                foreach (Match m in mc1)
+                {
+                    Match m0 = new Regex("(?<=href=\"[\\s]*?)([\\S]*?)(?=[\\s]*?\")").Match(m.Value);
+                    if (!m0.Success) m0 = new Regex("(?<=href=)([\\S]*?)(?=[\\s|>])").Match(m.Value);
+                    Match m1 = new Regex("(?<=>)([\\S\\s]*?)(?=</a>)").Match(m.Value);
+                    if (!m0.Success || !m1.Success) { content = content.Replace(m.Value, ""); continue; }
+                    string sHref = HtmlHref(m0.Value, url);
+                    //string s1 = ooData.sG10事物Get.Replace("^名称^",GetSqlParam(sHref));
+                    string s1 = ooData.sG10事物Get_p;
+                    List<SqlParameter> sqlparms = new List<SqlParameter>()
+                    {
+                        new SqlParameter() {ParameterName = "@名称", SqlDbType = SqlDbType.NVarChar, Value = GetSqlParam(sHref)}
+                    };
+                    //string s2 = ExecuteSQLResult(s1);
+                    string s2 = ExecuteSQLResult(s1, sqlparms);
+                    if (s2 != "")
+                        content = content.Replace(m.Value, "[链接[c" + int.Parse(s2).ToString("D6") + "[链接]" + m1.Value + "]链接]");
+                    else content = content.Replace(m.Value, m1.Value);
+                }
+
+                MatchCollection mc0 = new Regex("(<)([\\S\\s]*?)(>)").Matches(content);
+                foreach (Match m in mc0)
+                {
+                    if (m.Value == "<p>" || m.Value == "</p>" || m.Value == "<p class=\"artl\">" || m.Value == "<p class=\"artc\">"
+                        || m.Value == "<p class=\"artl\" style=\"text-align:center;\">" || m.Value == "<p style=\"text-align:center;\">"
+                        || m.Value == "<b>" || m.Value == "</b>" || m.Value == "<br/>" || m.Value == "<br>" || m.Value == "<br />"
+                        || m.Value == "<strong>" || m.Value == "</strong>"
+                        || m.Value == "</h1>" || m.Value == "</h2>" || m.Value == "</h3>" || m.Value == "</h4>" || m.Value == "</h5>"
+                        || m.Value == "<em>" || m.Value == "</em>" || m.Value == "<sup>" || m.Value == "</sup>") ;
+                    else if (m.Value.StartsWith("<h1")) { content = content.Replace(m.Value, "<h1 style=\"text-align:center;\">"); }
+                    else if (m.Value.StartsWith("<h2")) { content = content.Replace(m.Value, "<h2 style=\"text-align:center;\">"); }
+                    else if (m.Value.StartsWith("<h3")) { content = content.Replace(m.Value, "<h3 style=\"text-align:center;\">"); }
+                    else if (m.Value.StartsWith("<h4")) { content = content.Replace(m.Value, "<h4 style=\"text-align:center;\">"); }
+                    else if (m.Value.StartsWith("<h5")) { content = content.Replace(m.Value, "<h5 style=\"text-align:center;\">"); }
+                    else if (m.Value.StartsWith("<p")) content = content.Replace(m.Value, "<p>");
+                    else if (m.Value.StartsWith("<img") || m.Value.StartsWith("<IMG"))
+                    {
+                        Match m0 = new Regex("(?<=data-src=\")([\\S\\s]*?)(?=\")").Match(m.Value);
+                        if (!m0.Success) m0 = new Regex("(?<=src=\")([\\S\\s]*?)(?=\")").Match(m.Value);
+                        if (!m0.Success) m0 = new Regex("(?<=src=)([\\S]*?)(?=[\\s|>])").Match(m.Value);
+                        if (m0.Value.StartsWith("data:image/")) continue;
+                        Match m1 = new Regex("(?<=width=\")([\\S\\s]*?)(?=\")").Match(m.Value);
+                        string sHref = GetHtmlParam(HtmlHref(m0.Value, url));
+                        int w = 0; int.TryParse(m1.Value, out w);
+                        if (m0.Value == "/rmb/images/rmb_donation_200x200.jpg"
+                            || sHref == "https://www.rfa.org/++plone++rfa-resources/img/icon-zoom.png") content = content.Replace(m.Value, "");
+                        else
+                        {
+                            sHref = GetFile(sHref);
+                            if (w < 45 && w > 0) content = content.Replace(m.Value, "<img src=\"" + EnUrl2(sHref) + "\">");
+                            else content = content.Replace(m.Value, "<img class=\"artl\" src=\"" + EnUrl2(sHref) + "\">");
+                        }
+                    }
+                    else if (m.Value.StartsWith("<table")) content = content.Replace(m.Value, "<table>");
+                    else if (m.Value.StartsWith("<thead")) content = content.Replace(m.Value, "<thead>");
+                    else if (m.Value.StartsWith("<tr")) content = content.Replace(m.Value, "<tr>");
+                    else if (m.Value.StartsWith("<td"))
+                    {
+                        Match m0 = new Regex("( rowspan=\")([\\S\\s]*?)(\")").Match(m.Value);
+                        Match m1 = new Regex("( colspan=\")([\\S\\s]*?)(\")").Match(m.Value);
+                        content = content.Replace(m.Value, "<td" + m0.Value + m1.Value + ">");
+                    }
+                    else if (m.Value.StartsWith("<th"))
+                    {
+                        Match m0 = new Regex("( rowspan=\")([\\S\\s]*?)(\")").Match(m.Value);
+                        Match m1 = new Regex("( colspan=\")([\\S\\s]*?)(\")").Match(m.Value);
+                        content = content.Replace(m.Value, "<th" + m0.Value + m1.Value + ">");
+                    }
+                    else if (m.Value == "</th>" || m.Value == "</td>" || m.Value == "</tr>"
+                        || m.Value == "<tbody>" || m.Value == "</tbody>"
+                        || m.Value == "</thead>" || m.Value == "</table>") ;
+                    else if (m.Value.StartsWith("<sup")) content = content.Replace(m.Value, "<sup>");
+                    else if (m.Value == "<li>" || m.Value.StartsWith("<li ")) content = content.Replace(m.Value, "<li>");
+                    else if (m.Value == "</li>") ;
+                    else if (m.Value.StartsWith("<ol")) content = content.Replace(m.Value, "<ol>");
+                    else if (m.Value == "</ol>") ;
+                    else content = content.Replace(m.Value, "");
+                }
+
+                content = content.Replace("\t", "").Replace("\n\n", "\n").Replace("\n\n", "\n").Replace("\n\n", "\n")
+                    .Replace("\n", "\r\n").Replace("\r\r\n", "\r\n");
                 return content;
             }
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
             return "";
-            //try
-            //{
-            //    content = HtmlDel(content, "(<head)([\\S\\s]*?)(</head>)");
-            //    content = HtmlDel(content, "(<script)([\\S\\s]*?)(</script>)");
-            //    content = HtmlDel(content, "(<noscript)([\\S\\s]*?)(</noscript>)");
-            //    content = HtmlDel(content, "(<style)([\\S\\s]*?)(</style>)");
-            //    content = HtmlDel(content, "(<ins)([\\S\\s]*?)(</ins>)");
-
-            //    MatchCollection mc1 = new Regex("(<a )([\\S\\s]*?)(</a>)").Matches(content);
-            //    foreach (Match m in mc1)
-            //    {
-            //        Match m0 = new Regex("(?<=href=\"[\\s]*?)([\\S]*?)(?=[\\s]*?\")").Match(m.Value);
-            //        if (!m0.Success) m0 = new Regex("(?<=href=)([\\S]*?)(?=[\\s|>])").Match(m.Value);
-            //        Match m1 = new Regex("(?<=>)([\\S\\s]*?)(?=</a>)").Match(m.Value);
-            //        if (!m0.Success || !m1.Success) { content = content.Replace(m.Value, ""); continue; }
-            //        string sHref = HtmlHref(m0.Value, url);
-            //        string s1 = ooData.sG10事物Get.Replace("^名称^",GetSqlParam(sHref));
-            //        string s2 = ExecuteSQLResult(s1);
-            //        if (s2 != "")
-            //            content = content.Replace(m.Value, "[链接[c" + int.Parse(s2).ToString("D6") + "[链接]" + m1.Value + "]链接]");
-            //        else content = content.Replace(m.Value, m1.Value);
-            //    }
-
-            //    MatchCollection mc0 = new Regex("(<)([\\S\\s]*?)(>)").Matches(content);
-            //    foreach (Match m in mc0)
-            //    {
-            //        if (m.Value == "<p>" || m.Value == "</p>" || m.Value == "<p class=\"artl\">" || m.Value == "<p class=\"artc\">"
-            //            || m.Value == "<p class=\"artl\" style=\"text-align:center;\">" || m.Value == "<p style=\"text-align:center;\">"
-            //            || m.Value == "<b>" || m.Value == "</b>" || m.Value == "<br/>" || m.Value == "<br>" || m.Value == "<br />"
-            //            || m.Value == "<strong>" || m.Value == "</strong>"
-            //            || m.Value == "</h1>" || m.Value == "</h2>" || m.Value == "</h3>" || m.Value == "</h4>" || m.Value == "</h5>"
-            //            || m.Value == "<em>" || m.Value == "</em>" || m.Value == "<sup>" || m.Value == "</sup>") ;
-            //        else if (m.Value.StartsWith("<h1")) { content = content.Replace(m.Value, "<h1 style=\"text-align:center;\">"); }
-            //        else if (m.Value.StartsWith("<h2")) { content = content.Replace(m.Value, "<h2 style=\"text-align:center;\">"); }
-            //        else if (m.Value.StartsWith("<h3")) { content = content.Replace(m.Value, "<h3 style=\"text-align:center;\">"); }
-            //        else if (m.Value.StartsWith("<h4")) { content = content.Replace(m.Value, "<h4 style=\"text-align:center;\">"); }
-            //        else if (m.Value.StartsWith("<h5")) { content = content.Replace(m.Value, "<h5 style=\"text-align:center;\">"); }
-            //        else if (m.Value.StartsWith("<p")) content = content.Replace(m.Value, "<p>");
-            //        else if (m.Value.StartsWith("<img") || m.Value.StartsWith("<IMG"))
-            //        {
-            //            Match m0 = new Regex("(?<=data-src=\")([\\S\\s]*?)(?=\")").Match(m.Value);
-            //            if (!m0.Success) m0 = new Regex("(?<=src=\")([\\S\\s]*?)(?=\")").Match(m.Value);
-            //            if (!m0.Success) m0 = new Regex("(?<=src=)([\\S]*?)(?=[\\s|>])").Match(m.Value);
-            //            if (m0.Value.StartsWith("data:image/")) continue;
-            //            Match m1 = new Regex("(?<=width=\")([\\S\\s]*?)(?=\")").Match(m.Value);
-            //            string sHref = GetHtmlParam(HtmlHref(m0.Value, url));
-            //            int w = 0; int.TryParse(m1.Value, out w);
-            //            if (m0.Value == "/rmb/images/rmb_donation_200x200.jpg"
-            //                || sHref == "https://www.rfa.org/++plone++rfa-resources/img/icon-zoom.png") content = content.Replace(m.Value, "");
-            //            else
-            //            {
-            //                sHref = GetFile(sHref);
-            //                if (w < 45 && w > 0) content = content.Replace(m.Value, "<img src=\"" + EnUrl2(sHref) + "\">");
-            //                else content = content.Replace(m.Value, "<img class=\"artl\" src=\"" + EnUrl2(sHref) + "\">");
-            //            }
-            //        }
-            //        else if (m.Value.StartsWith("<table")) content = content.Replace(m.Value, "<table>");
-            //        else if (m.Value.StartsWith("<thead")) content = content.Replace(m.Value, "<thead>");
-            //        else if (m.Value.StartsWith("<tr")) content = content.Replace(m.Value, "<tr>");
-            //        else if (m.Value.StartsWith("<td"))
-            //        {
-            //            Match m0 = new Regex("( rowspan=\")([\\S\\s]*?)(\")").Match(m.Value);
-            //            Match m1 = new Regex("( colspan=\")([\\S\\s]*?)(\")").Match(m.Value);
-            //            content = content.Replace(m.Value, "<td" + m0.Value + m1.Value + ">");
-            //        }
-            //        else if (m.Value.StartsWith("<th"))
-            //        {
-            //            Match m0 = new Regex("( rowspan=\")([\\S\\s]*?)(\")").Match(m.Value);
-            //            Match m1 = new Regex("( colspan=\")([\\S\\s]*?)(\")").Match(m.Value);
-            //            content = content.Replace(m.Value, "<th" + m0.Value + m1.Value + ">");
-            //        }
-            //        else if (m.Value == "</th>" || m.Value == "</td>" || m.Value == "</tr>"
-            //            || m.Value == "<tbody>" || m.Value == "</tbody>"
-            //            || m.Value == "</thead>" || m.Value == "</table>") ;
-            //        else if (m.Value.StartsWith("<sup")) content = content.Replace(m.Value, "<sup>");
-            //        else if (m.Value == "<li>" || m.Value.StartsWith("<li ")) content = content.Replace(m.Value, "<li>");
-            //        else if (m.Value == "</li>") ;
-            //        else if (m.Value.StartsWith("<ol")) content = content.Replace(m.Value, "<ol>");
-            //        else if (m.Value == "</ol>") ;
-            //        else content = content.Replace(m.Value, "");
-            //    }
-
-            //    content = content.Replace("\t", "").Replace("\n\n", "\n").Replace("\n\n", "\n").Replace("\n\n", "\n")
-            //        .Replace("\n", "\r\n").Replace("\r\r\n", "\r\n");
-            //    return content;
-            //}
-            //catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
-            //return "";
         }
 
-        private static string HtmlHref(string sHref, string url)
+        static private string HtmlHref(string sHref, string url)
         {
             try
             {
@@ -1163,7 +1177,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return sHref;
         }
 
-        public static string UrlProc(string name)
+        static public string UrlProc(string name)
         {
             try
             {
@@ -1201,208 +1215,111 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
             return name;
         }
-
-        private static string GetHtmlParam(string name)
+        static private string GetHtmlParam(string name)
         {
             return name.Replace("<", "&lt;").Replace(">", "&gt;")
                 .Replace("'", "&#39;").Replace("\"", "&quot;");
         }
 
         // HtmlRsc
-        public static string EnUrl2(string name, bool bFirst = false)
+        static public string EnUrl2(string name, bool bFirst = false)
         {
             return name;
         }
 
-        public static string EnUrlSymbol(string url)
+        static public string EnUrlSymbol(string url)
         {
             return url.Replace(":", "%3A").Replace("/", "%2F").Replace("?", "%3F").Replace("&", "%26").Replace("#", "%23").Replace("=", "%3D");
         }
 
         // Utility
-        public static void Log(string co)
+        static public void Log(string co)
         {
             WriteErr(co);
         }
 
-        public static void Log(string name, string co)
+        static public void Log(string name, string co)
         {
             Log(name + ": " + co);
         }
 
-        private static string sRandomChars = "abcdefghijklmnopqrstuvwxyz";
-        private static string sRandomChars2 = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-        private static object oRandom = new object();
-
-        public static string GetRandom(int len = 12, bool upper = false)
+        static private string sRandomChars = "abcdefghijklmnopqrstuvwxyz";
+        static private string sRandomChars2 = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+        static private object oRandom = new object();
+        static public string GetRandom(int len = 12, bool upper = false)
         {
             string randomChars = upper ? sRandomChars2 : sRandomChars;
             string password = "";
             lock (oRandom) for (int i = 0; i < len; i++) password += randomChars[random.Next(randomChars.Length)];
             return password;
         }
-
-        public static string GetRandomNum()
+        static public string GetRandomNum()
         {
             string password = "";
             lock (oRandom) password = random.Next(10000000, 99999999).ToString();
             return password;
         }
-
-        //private static string GetTxt部分(string s摘要, string s链接, int i长度)
-        //{
-        //    if (s摘要.Length > i长度)
-        //    {
-        //        string s = "";
-        //        string[] ss = s摘要.Replace("\r\n", "\n").Replace("\r", "\n").Split('\n');
-        //        for (int i1 = 0; i1 < ss.Length; i1++)
-        //            if (s.Length + ss[i1].Length <= i长度 && !ss[i1].StartsWith("#{"))
-        //                s += (s == "" ? "" : "\n") + ss[i1];
-        //            else break;
-        //        s摘要 = s + "\n[链接[" + s链接 + "[链接][语言[全文]语言]]链接]";
-        //    }
-        //    return s摘要;
-        //}
-
-        //private static string GetString千万(string page)
-        //{
-        //    if (page.Length >= 5) return page.Substring(0, page.Length - 4) + "万";
-        //    else if (page.Length >= 4) return page.Substring(0, page.Length - 3) + "千";
-        //    else return page;
-        //}
-
-        //private static string GetString最长(string sIn, int iLen)
-        //{
-        //    if (sIn.Length <= iLen) return sIn;
-        //    return sIn.Substring(0, iLen - 1) + "…";
-        //}
-
-        //private static string Get一行(string sIn)
-        //{
-        //    return Get换行(sIn).Replace("\n", " ");
-        //}
-
-        private static string Get换行(string sIn)
+        static private string GetPictParam(string sPict, string sNumb)
         {
-            return sIn.Replace("\r\n", "\n").Replace("\r", "\n");
+            sPict = GetHtmlParam(sPict.Trim());
+            if (sPict != "")
+                return "><img class='imgh' src='" + EnUrl2(sPict.Replace("#P ", "").Replace(".jpg", "").Replace(".jpeg", "").Replace(".png", "")
+                    + ".K.jpg", true) + "'/>";
+            else return " style='background:hsla(" + int.Parse(sNumb) % 360 + ",75%,50%,0.5)'>";
         }
 
-        //private static string Get时距Param(DateTime dt)
-        //{
-        //    TimeSpan tsNow = DateTime.Now - dt;
-        //    if (tsNow.TotalMinutes < 1.0) return "[语言[刚刚]语言]";
-        //    if (tsNow.TotalHours < 1.0) return Math.Floor(tsNow.TotalMinutes) + "[语言[分前]语言]";
-        //    if (tsNow.TotalDays < 1.0) return Math.Floor(tsNow.TotalHours) + "[语言[时前]语言]";
-        //    return Math.Floor(tsNow.TotalDays) + "[语言[天前]语言]";
-        //}
+        static public string RSAEncrypt(string publickey, string content)
+        {
+            try
+            {
+                RSACryptoServiceProvider rsa = new RSACryptoServiceProvider();
+                rsa.FromXmlString(publickey);
+                byte[] cipherbytes = rsa.Encrypt(Encoding.UTF8.GetBytes(content), false);
+                return Convert.ToBase64String(cipherbytes);
+            }
+            catch { return ""; }
+        }
 
-        //private static string GetPict中(string sPict)
-        //{
-        //    return GetHtmlParam(sPict.Trim()).Replace("#P ", "").Replace(".jpg", "").Replace(".jpeg", "").Replace(".png", "") + ".M.jpg";
-        //}
-
-        //private static string GetPictParam(string sPict, string sNumb)
-        //{
-        //    sPict = GetHtmlParam(sPict.Trim());
-        //    if (sPict != "")
-        //        return "><img class='imgh' src='" + EnUrl2(sPict.Replace("#P ", "").Replace(".jpg", "").Replace(".jpeg", "").Replace(".png", "")
-        //            + ".K.jpg", true) + "'/>";
-        //    else return " style='background:hsla(" + int.Parse(sNumb) % 360 + ",75%,50%,0.5)'>";
-        //}
-
-        //private static string Get时间(string s)
-        //{
-        //    string r = "";
-        //    try
-        //    {
-        //        r = DateTime.Parse(s).ToString("yyyy-MM-dd HH:mm");
-        //    }
-        //    catch { }
-        //    return r;
-        //}
-
-        //public static string HtmlEn(string input, string password)
-        //{
-        //    string r = "";
-        //    if (input.Length == 0) return "";
-        //    r += string.Format("{0:x4}", ((int)input[0]));// + input.Length));
-        //    for (int i = 1; i < input.Length; i++) r += string.Format("{0:x4}", ((int)input[i]));// + (int)input[i - 1]);
-        //    return r;
-        //}
-
-        //public static string RSAEncrypt(string publickey, string content)
-        //{
-        //    try
-        //    {
-        //        RSACryptoServiceProvider rsa = new RSACryptoServiceProvider();
-        //        rsa.FromXmlString(publickey);
-        //        byte[] cipherbytes = rsa.Encrypt(Encoding.UTF8.GetBytes(content), false);
-        //        return Convert.ToBase64String(cipherbytes);
-        //    }
-        //    catch { return ""; }
-        //}
-
-        //public static string RSADecrypt(string privatekey, string content)
-        //{
-        //    try
-        //    {
-        //        RSACryptoServiceProvider rsa = new RSACryptoServiceProvider();
-        //        rsa.FromXmlString(privatekey);
-        //        byte[] cipherbytes = rsa.Decrypt(Convert.FromBase64String(content), false);
-        //        return Encoding.UTF8.GetString(cipherbytes);
-        //    }
-        //    catch { return ""; }
-        //}
+        static public string RSADecrypt(string privatekey, string content)
+        {
+            try
+            {
+                RSACryptoServiceProvider rsa = new RSACryptoServiceProvider();
+                rsa.FromXmlString(privatekey);
+                byte[] cipherbytes = rsa.Decrypt(Convert.FromBase64String(content), false);
+                return Encoding.UTF8.GetString(cipherbytes);
+            }
+            catch { return ""; }
+        }
 
         // Show
         private delegate void ShowMsgDelegate(string msg);
-
         private void ShowLabel(string msg)
         {
             labelMsg.Text = msg;
         }
-
-        //private void ShowLabelD(string msg)
-        //{
-        //    BeginInvoke(new ShowMsgDelegate(ShowLabel), new object[] { msg });
-        //}
-
         private void ShowMsg(string msg)
         {
             textBoxMsg.AppendText("[" + DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") + "] " + msg + "\r\n");
             textBoxMsg.ScrollToCaret();
         }
-
         private void ShowMsgD(string msg)
         {
             //WriteMsg(msg);
             BeginInvoke(new ShowMsgDelegate(ShowMsg), new object[] { msg });
         }
-
         private void ShowErr(string msg)
         {
             textBoxMsg.AppendText("[" + DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") + "] " + msg + "\r\n");
             textBoxMsg.ScrollToCaret();
         }
-
         private void ShowErrD(string msg)
         {
             //WriteErr(msg);
             BeginInvoke(new ShowMsgDelegate(ShowErr), new object[] { msg });
         }
 
-        //private static void WriteMsg(string message)
-        //{
-        //    try
-        //    {
-        //        StreamWriter swlog = File.AppendText(sAppName + "Msg.txt");
-        //        swlog.WriteLine("[" + DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") + "] " + message);
-        //        swlog.Close();
-        //    }
-        //    catch { }
-        //}
-
-        private static void WriteErr(string message)
+        static private void WriteErr(string message)
         {
             try
             {
@@ -1412,7 +1329,6 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             }
             catch { }
         }
-
         private void SaveFile(string message, string sFile)
         {
             try
@@ -1425,7 +1341,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
         }
 
         // Html Utility
-        public static string GetHtml(string sName, bool bFail = false)
+        static public string GetHtml(string sName, bool bFail = false)
         {
             string r = "";
             r = GetHtmlMethod("GET", sName, "", "", "", "");
@@ -1433,18 +1349,13 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return r;
         }
 
-        //public static string PostHtml(string sName, string sData)
-        //{
-        //    return GetHtmlMethod("POST", sName, sData, "", "", "");
-        //}
-
-        public static string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
+        static public string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
             string sCode)
         {
             return GetHtmlMethod(sMethod, sName, sData, sHeader, sReferer, sCode, "", out string c);
         }
 
-        public static string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
+        static public string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
             string sCode, string sContentType, out string sCookie)
         {
             sCookie = "";
@@ -1504,224 +1415,62 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return r;
         }
 
-        //public static string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
-        //    string sContentType, string sHost, string sCode)
-        //{
-        //    string r = "";
-        //    HttpWebRequest request = null;
-        //    HttpWebResponse response = null;
-        //    Stream requestStream = null;
-        //    Stream responseStream = null;
-        //    StreamReader streamReader = null;
-        //    try
-        //    {
-        //        request = (HttpWebRequest)WebRequest.Create(sName);
-        //        request.UserAgent = sUserAgent;
-        //        if (sReferer != "") request.Referer = sReferer;
-        //        if (sHost != "") request.Host = sHost;
-        //        if (sContentType != "") request.ContentType = sContentType;
-        //        else if (sMethod == "POST") request.ContentType = "application/x-www-form-urlencoded";
-        //        request.Method = sMethod;
-        //        if (sHeader != "")
-        //        {
-        //            string[] ss0 = sHeader.Split(',');
-        //            foreach (string s in ss0)
-        //            {
-        //                string s0 = s.Substring(0, s.IndexOf(':'));
-        //                string s1 = s.Substring(s.IndexOf(':') + 1);
-        //                request.Headers[s0.Trim()] = s1.Trim();
-        //            }
-        //        }
-        //        if (sData != "")
-        //        {
-        //            byte[] postData = Encoding.UTF8.GetBytes(sData);
-        //            request.ContentLength = postData.Length;
-        //            requestStream = request.GetRequestStream();
-        //            requestStream.Write(postData, 0, postData.Length);
-        //            requestStream.Close();
-        //        }
-        //        response = (HttpWebResponse)request.GetResponse();
-        //        if (response.StatusCode == HttpStatusCode.OK)
-        //        {
-        //            responseStream = response.GetResponseStream();
-        //            if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("gzip"))
-        //                responseStream = new GZipStream(responseStream, CompressionMode.Decompress);
-        //            else if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("deflate"))
-        //                responseStream = new DeflateStream(responseStream, CompressionMode.Decompress);
-        //            streamReader = new StreamReader(responseStream, Encoding.GetEncoding(sCode == "" ? "utf-8" : sCode));
-        //            r = streamReader.ReadToEnd();
-        //        }
-        //    }
-        //    catch (WebException ex)
-        //    {
-        //    }
-        //    catch (Exception ex) { }
-        //    finally
-        //    {
-        //        if (request != null) request = null;
-        //        if (response != null) { response.Close(); response = null; }
-        //        if (requestStream != null) requestStream.Close();
-        //        if (responseStream != null) responseStream.Close();
-        //        if (streamReader != null) streamReader.Close();
-        //    }
-        //    return r;
-        //}
+        private static bool AlwaysGoodCertificate(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors policyErrors)
+        {
+            return true;
+        }
 
-        //public static string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer,
-        //    string sContentType, string sCode)
-        //{
-        //    string r = "";
-        //    HttpWebRequest request = null;
-        //    HttpWebResponse response = null;
-        //    Stream requestStream = null;
-        //    Stream responseStream = null;
-        //    StreamReader streamReader = null;
-        //    try
-        //    {
-        //        request = (HttpWebRequest)WebRequest.Create(sName);
-        //        request.UserAgent = sUserAgent;
-        //        if (sReferer != "") request.Referer = sReferer;
-        //        if (sContentType != "") request.ContentType = sContentType;
-        //        else if (sMethod == "POST") request.ContentType = "application/x-www-form-urlencoded";
-        //        request.Method = sMethod;
-        //        if (sHeader != "")
-        //        {
-        //            string[] ss0 = sHeader.Split(',');
-        //            foreach (string s in ss0)
-        //            {
-        //                string s0 = s.Substring(0, s.IndexOf(':'));
-        //                string s1 = s.Substring(s.IndexOf(':') + 1);
-        //                request.Headers[s0.Trim()] = s1.Trim();
-        //            }
-        //        }
-        //        if (sData != "")
-        //        {
-        //            byte[] postData = Encoding.UTF8.GetBytes(sData);
-        //            request.ContentLength = postData.Length;
-        //            requestStream = request.GetRequestStream();
-        //            requestStream.Write(postData, 0, postData.Length);
-        //            requestStream.Close();
-        //        }
-        //        response = (HttpWebResponse)request.GetResponse();
-        //        if (response.StatusCode == HttpStatusCode.OK)
-        //        {
-        //            responseStream = response.GetResponseStream();
-        //            if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("gzip"))
-        //                responseStream = new GZipStream(responseStream, CompressionMode.Decompress);
-        //            else if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("deflate"))
-        //                responseStream = new DeflateStream(responseStream, CompressionMode.Decompress);
-        //            streamReader = new StreamReader(responseStream, Encoding.GetEncoding(sCode == "" ? "utf-8" : sCode));
-        //            r = streamReader.ReadToEnd();
-        //        }
-        //    }
-        //    catch (WebException ex)
-        //    {
-        //    }
-        //    catch (Exception ex) { }
-        //    finally
-        //    {
-        //        if (request != null) request = null;
-        //        if (response != null) { response.Close(); response = null; }
-        //        if (requestStream != null) requestStream.Close();
-        //        if (responseStream != null) responseStream.Close();
-        //        if (streamReader != null) streamReader.Close();
-        //    }
-        //    return r;
-        //}
+        public async Task<string> DownloadFile(string name, string host, string referer, string sFileName)
+        {
+            // validation
+            HttpClient _httpClient = new HttpClient();
+            using (var stream = await _httpClient.GetStreamAsync(name))
+            {
+                using (var fileStream = new FileStream(sFileName, FileMode.CreateNew))
+                {
+                    await stream.CopyToAsync(fileStream);
+                    return fileStream.Name;
+                }
+            }
 
-        //public static string GetHtmlMethod(string sMethod, string sName, string sData, string sHeader, string sReferer)
-        //{
-        //    string r = "";
-        //    HttpWebRequest request = null;
-        //    HttpWebResponse response = null;
-        //    Stream requestStream = null;
-        //    Stream responseStream = null;
-        //    StreamReader streamReader = null;
-        //    try
-        //    {
-        //        request = (HttpWebRequest)WebRequest.Create(sName);
-        //        request.UserAgent = sUserAgent;
-        //        if (sReferer != "") request.Referer = sReferer;
-        //        else if (sMethod == "POST") request.ContentType = "application/x-www-form-urlencoded";
-        //        request.Method = sMethod;
-        //        if (sHeader != "")
-        //        {
-        //            string[] ss0 = sHeader.Split(',');
-        //            foreach (string s in ss0)
-        //            {
-        //                string s0 = s.Substring(0, s.IndexOf(':'));
-        //                string s1 = s.Substring(s.IndexOf(':') + 1);
-        //                request.Headers[s0.Trim()] = s1.Trim();
-        //            }
-        //        }
-        //        if (sData != "")
-        //        {
-        //            byte[] postData = Encoding.UTF8.GetBytes(sData);
-        //            request.ContentLength = postData.Length;
-        //            requestStream = request.GetRequestStream();
-        //            requestStream.Write(postData, 0, postData.Length);
-        //            requestStream.Close();
-        //        }
-        //        response = (HttpWebResponse)request.GetResponse();
-        //        if (response.StatusCode == HttpStatusCode.OK)
-        //        {
-        //            responseStream = response.GetResponseStream();
-        //            if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("gzip"))
-        //                responseStream = new GZipStream(responseStream, CompressionMode.Decompress);
-        //            else if (response.ContentEncoding != null && response.ContentEncoding.ToLower().Contains("deflate"))
-        //                responseStream = new DeflateStream(responseStream, CompressionMode.Decompress);
-        //            streamReader = new StreamReader(responseStream, Encoding.GetEncoding("utf-8"));
-        //            r = streamReader.ReadToEnd();
-        //        }
-        //    }
-        //    catch (WebException ex)
-        //    {
-        //    }
-        //    catch (Exception ex) { }
-        //    finally
-        //    {
-        //        if (request != null) request = null;
-        //        if (response != null) { response.Close(); response = null; }
-        //        if (requestStream != null) requestStream.Close();
-        //        if (responseStream != null) responseStream.Close();
-        //        if (streamReader != null) streamReader.Close();
-        //    }
-        //    return r;
-        //}
+            //var fileInfo = new FileInfo(sFileName);
 
-        //public static string GetHtmlMethodOrigin(string sName)
-        //{
-        //    string r = "";
-        //    HttpWebRequest request = null;
-        //    HttpWebResponse response = null;
-        //    try
-        //    {
-        //        if (!sName.StartsWith("http://") && !sName.StartsWith("https://")) return "";
-        //        request = (HttpWebRequest)WebRequest.Create(sName);
-        //        request.UserAgent = sUserAgent;
-        //        request.Timeout = request.ReadWriteTimeout = 10 * 1000;
-        //        request.AllowAutoRedirect = false;
-        //        response = (HttpWebResponse)request.GetResponse();
-        //        r = response.Headers["Location"];
-        //    }
-        //    catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message); }
-        //    finally
-        //    {
-        //        if (request != null) request = null;
-        //        if (response != null) { response.Close(); response = null; }
-        //    }
-        //    return r;
-        //}
+            //Task<HttpResponseMessage> response = _httpClient.GetAsync(name);
+            //response.EnsureSuccessStatusCode();
+            //Stream ms = await response.Content.ReadAsStreamAsync();
+            //FileStream fs = File.Create(fileInfo.FullName);
+            //ms.Seek(0, SeekOrigin.Begin);
+            //ms.CopyTo(fs);
 
+            //return fileInfo.FullName;
+
+
+
+
+        }
         public bool DownloadHtml(string name, string host, string referer, string sFileName)
         {
             HttpWebRequest request2 = null;
             HttpWebResponse response2 = null;
             try
             {
+                //ServicePointManager.Expect100Continue = true;
+                //ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls12 | SecurityProtocolType.Ssl3;
+
+                //ServicePointManager.ServerCertificateValidationCallback = delegate { return true; };
+
+                ServicePointManager.SecurityProtocol = SecurityProtocolType.Ssl3;
+                ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls12;
+                ServicePointManager.ServerCertificateValidationCallback += new RemoteCertificateValidationCallback(AlwaysGoodCertificate);
+
+                //request2 = (HttpWebRequest)WebRequest.Create("https://a57.foxnews.com/static.foxnews.com/foxnews.com/content/uploads/2022/09/1024/576/GettyImages-1402371449.jpg?tl=1&ve=1");
+                //response2 = (HttpWebResponse)request2.GetResponse();
+
                 request2 = (HttpWebRequest)WebRequest.Create(name);
+                request2.ServerCertificateValidationCallback += (sender, certificate, chain, errors) => { return CertificateValidationCallBack(sender, certificate, chain, errors); };
                 if (host != "") request2.Host = host;
                 if (referer != "") request2.Referer = referer;
+                request2.UserAgent = sUserAgent;
                 response2 = (HttpWebResponse)request2.GetResponse();
                 if (response2.StatusCode != HttpStatusCode.OK)
                 {
@@ -1740,27 +1489,41 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
                         return true;
                     }
                 }
-                ShowMsgD("Downloading: " + sFileName);
-
-                byte[] buffer = new byte[8 * 1024];
-                Stream outStream = File.Create(sFileName);
-                Stream inStream = response2.GetResponseStream();
-                long length = response2.ContentLength;
-                long total = 0;
-                int l = 0;
-                while ((l = inStream.Read(buffer, 0, buffer.Length)) > 0)
+                if (isMimeAcceptable(response2.ContentType))
                 {
-                    total += l;
-                    int progress = (int)(((float)total / length) * 100);
-                    BeginInvoke(new ShowMsgDelegate(ShowLabel), new object[] { progress + " %" });
-                    outStream.Write(buffer, 0, l);
+                    //L4
+                    ShowMsgD("Downloading: " + sFileName);
+
+                    byte[] buffer = new byte[8 * 1024];
+                    Stream outStream = File.Create(sFileName);
+                    Stream inStream = response2.GetResponseStream();
+                    long length = response2.ContentLength;
+                    long total = 0;
+                    int l = 0;
+                    while ((l = inStream.Read(buffer, 0, buffer.Length)) > 0)
+                    {
+                        total += l;
+                        int progress = (int)(((float)total / length) * 100);
+                        BeginInvoke(new ShowMsgDelegate(ShowLabel), new object[] { progress + " %" });
+                        outStream.Write(buffer, 0, l);
+                        //if (isMimeAcceptable(GetMimeType(buffer))) //screening MIME types
+                        //{
+                        //    outStream.Write(buffer, 0, l);
+                        //}
+                        //else
+                        //{
+                        //    outStream.Close();
+                        //    return false;
+                        //}
+                    }
+                    outStream.Close();
+                    if (response2.ContentLength != -1 && length != total) { }
+                    //BeginInvoke(new ShowMsgDelegate(ShowLabel), new object[] { "Done" });
+                    if (response2 != null) { response2.Close(); response2 = null; }
+                    if (request2 != null) request2 = null;
+                    return true;
                 }
-                outStream.Close();
-                if (response2.ContentLength != -1 && length != total) { }
-                //BeginInvoke(new ShowMsgDelegate(ShowLabel), new object[] { "Done" });
-                if (response2 != null) { response2.Close(); response2 = null; }
-                if (request2 != null) request2 = null;
-                return true;
+
             }
             catch (Exception ex) { }
             if (response2 != null) { response2.Close(); response2 = null; }
@@ -1768,32 +1531,23 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return false;
         }
 
-        //public bool CheckHtml(string sName)
-        //{
-        //    HttpWebRequest request = null;
-        //    HttpWebResponse response = null;
-        //    bool r = false;
-        //    try
-        //    {
-        //        if (!sName.StartsWith("http://") && !sName.StartsWith("https://")) return r;
-        //        request = (HttpWebRequest)WebRequest.Create(sName);
-        //        request.UserAgent = sUserAgent;
-        //        request.Timeout = request.ReadWriteTimeout = 30 * 1000;
-        //        response = (HttpWebResponse)request.GetResponse();
-        //        if (response.StatusCode == HttpStatusCode.OK) r = true;
-        //        else ShowMsgD(response.StatusCode + "");
-        //    }
-        //    catch (Exception ex) { ShowMsgD(ex.Message); }
-        //    finally
-        //    {
-        //        if (response != null) { response.Close(); response = null; }
-        //        if (request != null) request = null;
-        //    }
-        //    return r;
-        //}
-
         // SQL Utility
-        public static string ExecuteSQLResult(string s)
+        public string ExecuteSQLResult(string s, List<SqlParameter> sps)
+        {
+            string r = "";
+            try
+            {
+                DataSet ds = new DataSet();
+                //ExecuteSQL(s, ds);
+                ExecuteSQL(s, ds, sps);
+                if (ds.Tables[0].Rows.Count > 0) r = ds.Tables[0].Rows[0][0].ToString();
+                ds.Clear();
+                ds = null;
+            }
+            catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message + " " + s); }
+            return r;
+        }
+        public string ExecuteSQLResult(string s)
         {
             string r = "";
             try
@@ -1808,17 +1562,52 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return r;
         }
 
-        public static bool ExecuteSQL(string s)
-        {
-            return ExecuteSQL(s, null);
-        }
-
-        public static bool ExecuteSQL(string s, DataSet ds)
+        public bool ExecuteSQL(string s, DataSet ds, List<SqlParameter> sps)
         {
             string tbname = "result";
             try
             {
-                string connTusiStr = ConfigurationManager.AppSettings["DB"];
+                string connTusiStr = tbxDbConnString.Text; //ConfigurationManager.AppSettings["DB"];
+                SqlConnection conn = new SqlConnection(connTusiStr);
+                conn.Open();
+                if (ds == null)
+                {
+                    SqlCommand sqlCmd = new SqlCommand(s, conn);
+                    foreach (var sp in sps)
+                    {
+                        sqlCmd.Parameters.Add(sp);
+                    }
+
+                    object o = sqlCmd.ExecuteScalar();
+                }
+                else
+                {
+                    SqlDataAdapter da = new SqlDataAdapter(s, conn);
+                    foreach (var sp in sps)
+                    {
+                        da.SelectCommand.Parameters.Add(sp);
+                    }
+
+                    da.Fill(ds, tbname);
+                }
+                conn.Close();
+                return true;
+            }
+            catch (Exception ex) { Log(MethodBase.GetCurrentMethod().Name + ": " + ex.Message + " " + s); }
+            return false;
+        }
+
+        public bool ExecuteSQL(string s)
+        {
+            return ExecuteSQL(s, null);
+        }
+
+        public bool ExecuteSQL(string s, DataSet ds)
+        {
+            string tbname = "result";
+            try
+            {
+                string connTusiStr = tbxDbConnString.Text; //ConfigurationManager.AppSettings["DB"];
                 SqlConnection conn = new SqlConnection(connTusiStr);
                 conn.Open();
                 if (ds == null)
@@ -1838,7 +1627,7 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
             return false;
         }
 
-        public static string GetSqlParam(string s)
+        static public string GetSqlParam(string s)
         {
             return s.Replace("'", "''");
         }
@@ -1893,6 +1682,820 @@ aws s3 sync C:\oLink\File s3://^S3Bucket^/File --acl public-read --region eu-wes
                 MessageBox.Show("The certificate is invalid.");
                 return false;
             }
+        }
+
+        private bool loginawsiam(string bucketname, string accesskeyid, string seckey)
+        {
+            using (AmazonRDSClient rdsClient = new AmazonRDSClient(RegionEndpoint.USEast1))
+            {
+
+            }
+
+            return false;
+        }
+        private bool verifys3bucket(string bucketname, string accesskeyid, string seckey)
+        {
+            AmazonS3Client s3Client = new AmazonS3Client(accesskeyid, seckey);
+            ListBucketsResponse buckets = s3Client.ListBuckets();
+            foreach (var bucket in buckets.Buckets)
+            {
+                if (bucket.BucketName.ToUpper() == bucketname.ToUpper())
+                    return true;
+            }
+
+            return false;
+        }
+
+        private void OrganizeLocalFolders()
+        {
+            string localfiledirStr = ConfigurationManager.AppSettings["localfiledir"];
+            string localsitedirStr = ConfigurationManager.AppSettings["localsitedir"];
+            string appdirStr = Application.StartupPath;
+
+            localfiledirStr = appdirStr + "\\" + localfiledirStr;
+            localsitedirStr = appdirStr + "\\" + localsitedirStr;
+            if (!Directory.Exists(localsitedirStr))
+            {
+                DirectoryInfo siteDi = Directory.CreateDirectory(localsitedirStr);
+
+                if (siteDi is null)
+                {
+                    ShowMsg(localsitedirStr + " not exists.");
+                }
+            }
+
+            if (!Directory.Exists(localfiledirStr))
+            {
+                DirectoryInfo siteDi = Directory.CreateDirectory(localfiledirStr);
+
+                if (siteDi is null)
+                {
+                    ShowMsg(localfiledirStr + " not exists.");
+                }
+            }
+        }
+        private void ClearDangerousFilesinLocalFolders()
+        {
+            //beginning byte[] pattern determins file/MIME type, 8.2022
+            string localfiledirStr = ConfigurationManager.AppSettings["localfiledir"];
+            string localsitedirStr = ConfigurationManager.AppSettings["localsitedir"];
+            if (localfiledirStr == null) localfiledirStr = "File";
+            if (localsitedirStr == null) localsitedirStr = "Site";
+
+            OrganizeLocalFolders();
+
+            string appdirStr = Application.StartupPath;
+
+            localfiledirStr = appdirStr + "\\" + localfiledirStr;
+            localsitedirStr = appdirStr + "\\" + localsitedirStr;
+            ClearDangerousFilesinLocalFolder(localfiledirStr);
+            ClearDangerousFilesinLocalFolder(localsitedirStr);
+        }
+        private void ClearDangerousFilesinLocalFolder(string foldername)
+        {
+            if (!Directory.Exists(foldername))
+            {
+                ShowMsgD("Directory Not Found: " + foldername);
+                return;
+            }
+
+            string fileext_checked = "";
+            string[] filePaths = Directory.GetFiles(foldername);
+            for (int i = 0; i < filePaths.Length; i++)
+            {
+                if (File.Exists(filePaths[i]))
+                {
+                    string fileext = Path.GetExtension(filePaths[i]);
+                    if (fileext.ToLower() == fileext_checked.ToLower()) continue;
+
+                    byte[] file = File.ReadAllBytes(filePaths[i]);
+                    //binary file mime check: based on a mime and byte match library
+                    string mime = GetMimeType(file, filePaths[i]);
+                    if (!isMimeAcceptable(mime))
+                        mime = getMimeFromFile(filePaths[i]);   //based on the query in to "urlmon.dll" in windows system
+                    if (isMimeAcceptable(mime))
+                    {
+                        ShowMsgD("Acceptable Type (" + mime + "): " + Path.GetFileName(filePaths[i]));
+                        fileext_checked = fileext;
+                        continue;
+                    }
+
+                    //the dll should do the work most likely
+                    if (isInFileTypeList(fileext))
+                    {
+                        ShowMsgD("Acceptable Type (" + mime + "): " + Path.GetFileName(filePaths[i]));
+                        fileext_checked = fileext;
+                        continue;
+                    }
+
+                    FileAttributes fas = File.GetAttributes(filePaths[i]);
+                    ShowMsgD("Delete (File Type Error " + mime + "): " + Path.GetFileName(filePaths[i]));
+                    File.Delete(filePaths[i]);
+
+                }
+            }
+
+        }
+
+        // !isInFileTypeList(Path.GetExtension(filePaths[i]))
+        private bool isInFileTypeList(string fileext)
+        {
+            Microsoft.Win32.RegistryKey regKey = Microsoft.Win32.Registry.ClassesRoot.OpenSubKey(fileext.ToLower());
+            if (regKey != null && regKey.GetValue("Content Type") != null)
+            {
+                string mimeType;
+                mimeType = regKey.GetValue("Content Type").ToString();
+                //Console.WriteLine(mimeType);
+
+                return isMimeAcceptable(mimeType);
+            }
+            return false;
+        }
+        private void textBoxS3Id_TextChanged(object sender, EventArgs e)
+        {
+            //when start changing credentials, enable bucket entry
+            //textBoxS3Bucket.Enabled = true;
+        }
+
+        private void textBoxS3Key_TextChanged(object sender, EventArgs e)
+        {
+            //textBoxS3Bucket.Enabled = true;
+        }
+
+        private bool isMimeAcceptable(string mimestring)
+        {
+            if (mimestring.StartsWith("application/") && mimestring.Contains("javascript")
+                || mimestring.StartsWith("image/")
+                || mimestring.StartsWith("audio/")
+                || mimestring.StartsWith("video/")
+                || mimestring.StartsWith("application/octet-stream") && (mimestring.Contains("javascript"))
+                || mimestring.StartsWith("text/plain")
+                || mimestring.StartsWith("text/css")
+                || mimestring.StartsWith("text/html")
+                ) return true;
+            return false;
+        }
+        private string GetMimeType(byte[] filebytes)
+        {
+
+            string mime = "application/octet-stream"; //DEFAULT UNKNOWN MIME TYPE
+
+            //Get the MIME Type
+            if (filebytes.Take(2).SequenceEqual(BMP))
+            {
+                mime = "image/bmp";
+            }
+            else if (filebytes.Take(8).SequenceEqual(DOC))
+            {
+                mime = "application/msword";
+            }
+            else if (filebytes.Take(2).SequenceEqual(EXE_DLL))
+            {
+                mime = "application/x-msdownload"; //both use same mime type
+            }
+            else if (filebytes.Take(4).SequenceEqual(GIF))
+            {
+                mime = "image/gif";
+            }
+            else if (filebytes.Take(4).SequenceEqual(ICO))
+            {
+                mime = "image/x-icon";
+            }
+            else if (filebytes.Take(3).SequenceEqual(JPG))
+            {
+                mime = "image/jpeg";
+            }
+            else if (filebytes.Take(3).SequenceEqual(MP3))
+            {
+                mime = "audio/mpeg";
+            }
+            else if (filebytes.Take(14).SequenceEqual(OGG))
+            {
+                mime = "video/ogg";
+            }
+            else if (filebytes.Take(7).SequenceEqual(PDF))
+            {
+                mime = "application/pdf";
+            }
+            else if (filebytes.Take(16).SequenceEqual(PNG))
+            {
+                mime = "image/png";
+            }
+            else if (filebytes.Take(7).SequenceEqual(RAR))
+            {
+                mime = "application/x-rar-compressed";
+            }
+            else if (filebytes.Take(3).SequenceEqual(SWF))
+            {
+                mime = "application/x-shockwave-flash";
+            }
+            else if (filebytes.Take(4).SequenceEqual(TIFF))
+            {
+                mime = "image/tiff";
+            }
+            else if (filebytes.Take(11).SequenceEqual(TORRENT))
+            {
+                mime = "application/x-bittorrent";
+            }
+            else if (filebytes.Take(5).SequenceEqual(TTF))
+            {
+                mime = "application/x-font-ttf";
+            }
+            else if (filebytes.Take(4).SequenceEqual(WAV_AVI))
+            {
+                mime = "media/avi-wav";//extension == ".AVI" ? "video/x-msvideo" : "audio/x-wav";
+            }
+            else if (filebytes.Take(16).SequenceEqual(WMV_WMA))
+            {
+                mime = "media/wma-=wmv";// extension == ".WMA" ? "audio/x-ms-wma" : "video/x-ms-wmv";
+            }
+            else if (filebytes.Take(4).SequenceEqual(ZIP_DOCX))
+            {
+                mime = "application/word-zip";
+                //extension == ".DOCX" ? "application/vnd.openxmlformats-officedocument.wordprocessingml.document" : "application/x-zip-compressed";
+            }
+
+            return mime;
+        }
+        private string GetMimeType(byte[] file, string fileName)
+        {
+
+            string mime = "application/octet-stream"; //DEFAULT UNKNOWN MIME TYPE
+
+            //Ensure that the filename isn't empty or null
+            if (string.IsNullOrWhiteSpace(fileName))
+            {
+                return mime;
+            }
+
+            //Get the file extension
+            string extension = Path.GetExtension(fileName) == null
+                                   ? string.Empty
+                                   : Path.GetExtension(fileName).ToUpper();
+
+            //Get the MIME Type
+            if (file.Take(2).SequenceEqual(BMP))
+            {
+                mime = "image/bmp";
+            }
+            else if (file.Take(8).SequenceEqual(DOC))
+            {
+                mime = "application/msword";
+            }
+            else if (file.Take(2).SequenceEqual(EXE_DLL))
+            {
+                mime = "application/x-msdownload"; //both use same mime type
+            }
+            else if (file.Take(4).SequenceEqual(GIF))
+            {
+                mime = "image/gif";
+            }
+            else if (file.Take(4).SequenceEqual(ICO))
+            {
+                mime = "image/x-icon";
+            }
+            else if (file.Take(3).SequenceEqual(JPG))
+            {
+                mime = "image/jpeg";
+            }
+            else if (file.Take(3).SequenceEqual(MP3))
+            {
+                mime = "audio/mpeg";
+            }
+            else if (file.Take(14).SequenceEqual(OGG))
+            {
+                if (extension == ".OGX")
+                {
+                    mime = "application/ogg";
+                }
+                else if (extension == ".OGA")
+                {
+                    mime = "audio/ogg";
+                }
+                else
+                {
+                    mime = "video/ogg";
+                }
+            }
+            else if (file.Take(7).SequenceEqual(PDF))
+            {
+                mime = "application/pdf";
+            }
+            else if (file.Take(16).SequenceEqual(PNG))
+            {
+                mime = "image/png";
+            }
+            else if (file.Take(7).SequenceEqual(RAR))
+            {
+                mime = "application/x-rar-compressed";
+            }
+            else if (file.Take(3).SequenceEqual(SWF))
+            {
+                mime = "application/x-shockwave-flash";
+            }
+            else if (file.Take(4).SequenceEqual(TIFF))
+            {
+                mime = "image/tiff";
+            }
+            else if (file.Take(11).SequenceEqual(TORRENT))
+            {
+                mime = "application/x-bittorrent";
+            }
+            else if (file.Take(5).SequenceEqual(TTF))
+            {
+                mime = "application/x-font-ttf";
+            }
+            else if (file.Take(4).SequenceEqual(WAV_AVI))
+            {
+                mime = extension == ".AVI" ? "video/x-msvideo" : "audio/x-wav";
+            }
+            else if (file.Take(16).SequenceEqual(WMV_WMA))
+            {
+                mime = extension == ".WMA" ? "audio/x-ms-wma" : "video/x-ms-wmv";
+            }
+            else if (file.Take(4).SequenceEqual(ZIP_DOCX))
+            {
+                mime = extension == ".DOCX" ? "application/vnd.openxmlformats-officedocument.wordprocessingml.document" : "application/x-zip-compressed";
+            }
+
+            return mime;
+        }
+        [DllImport(@"urlmon.dll", CharSet = CharSet.Auto)]
+        private extern static System.UInt32 FindMimeFromData(
+        System.UInt32 pBC,
+        [MarshalAs(UnmanagedType.LPStr)] System.String pwzUrl,
+        [MarshalAs(UnmanagedType.LPArray)] byte[] pBuffer,
+        System.UInt32 cbSize,
+        [MarshalAs(UnmanagedType.LPStr)] System.String pwzMimeProposed,
+        System.UInt32 dwMimeFlags,
+        out System.UInt32 ppwzMimeOut,
+        System.UInt32 dwReserverd
+    );
+
+        public static string getMimeFromFile(string filename)
+        {
+            if (!File.Exists(filename))
+                throw new FileNotFoundException(filename + " not found");
+
+            byte[] buffer = new byte[256];
+            using (FileStream fs = new FileStream(filename, FileMode.Open))
+            {
+                if (fs.Length >= 256)
+                    fs.Read(buffer, 0, 256);
+                else
+                    fs.Read(buffer, 0, (int)fs.Length);
+            }
+            try
+            {
+                System.UInt32 mimetype;
+                FindMimeFromData(0, null, buffer, 256, null, 0, out mimetype, 0);
+                System.IntPtr mimeTypePtr = new IntPtr(mimetype);
+                string mime = Marshal.PtrToStringUni(mimeTypePtr);
+                Marshal.FreeCoTaskMem(mimeTypePtr);
+                return mime;
+            }
+            catch (Exception e)
+            {
+                return "unknown/unknown";
+            }
+        }
+
+        private void listprofilenames()
+        {
+            List<CredentialProfile> aws_credlist = loadProfileFromCredFile();
+
+            if (aws_credlist.Count > 0)
+            {
+                List<string> profnames = new List<string>();
+                foreach (var pf in aws_credlist)
+                {
+                    profnames.Add(pf.Name);
+                }
+
+                cbxProfiles.DataSource = profnames;
+                btnSaveCred.Text = "Use Profile";
+                EnableProfileEntries(false);
+                //cbxProfiles.SelectedIndex = -1;
+                return;
+            }
+
+            btnSaveCred.Text = "Create Profile";
+            EnableProfileEntries(true);
+            ShowhideDataTabs(false);
+        }
+        private void ShowhideDataTabs(bool bshow)
+        {
+            if (!bshow)
+            {
+                tabEntries.TabPages.Remove(tabPage1);
+                tabEntries.TabPages.Remove(tabPage2);
+                tabEntries.TabPages.Remove(tabPage3);
+                tabEntries.TabPages.Remove(tabPage4);
+            }
+            else
+            {
+                tabEntries.TabPages.Add(tabPage1);
+                tabEntries.TabPages.Add(tabPage2);
+                tabEntries.TabPages.Add(tabPage3);
+                tabEntries.TabPages.Add(tabPage4);
+            }
+        }
+
+        private void ShowhideCredsTab(bool bshow)
+        {
+            if (!bshow)
+            {
+                tabEntries.TabPages.Remove(tpIAM);
+            }
+            else
+            {
+                tabEntries.TabPages.Add(tpIAM);
+            }
+        }
+        private void EnableProfileEntries(bool benable)
+        {
+            tbxProfName.Enabled = benable;
+            tbxAccKey.Enabled = benable;
+            tbxSecKey.Enabled = benable;
+            tbxDbPwdP.Enabled = benable;
+            tbxDbUserP.Enabled = benable;
+        }
+
+        private void EnableDbLogins(bool benable)
+        {
+            tbxDbPwdP.Enabled = benable;
+            tbxDbUserP.Enabled = benable;
+        }
+
+        private void EnableEntryTabs(bool benable)
+        {
+            for (int i = 0; i < this.tabEntries.TabCount; i++)
+            {
+                if (tabEntries.TabPages[i].Name != "tpIAM")
+                {
+                    tabEntries.TabPages[i].Size = Size.Empty;
+                }
+            }
+        }
+
+        private string protectText(string texttoencrypt)
+        {
+            var sharedFile = new SharedCredentialsFile();
+            if (!File.Exists(sharedFile.FilePath)) return "";
+
+
+            string readText = File.ReadAllText(sharedFile.FilePath);
+
+
+            FileStream fStream = new FileStream(sharedFile.FilePath + encrypt_fileext, FileMode.OpenOrCreate, FileAccess.Read, FileShare.ReadWrite);
+
+            int length = 0;
+            byte[] toEncrypt = UnicodeEncoding.ASCII.GetBytes(readText);
+            byte[] entropy = CreateRandomEntropy();
+            byte[] encrptedData = ProtectedData.Protect(toEncrypt, entropy, DataProtectionScope.CurrentUser);
+            // Write the encrypted data to a stream.
+            if (fStream.CanWrite && encrptedData != null)
+            {
+                fStream.Write(encrptedData, 0, encrptedData.Length);
+
+                length = encrptedData.Length;
+            }
+
+            fStream.Close();
+            return "";
+        }
+
+        private string saveCredFile()
+        {
+            var sharedFile = new SharedCredentialsFile();
+            if (!File.Exists(sharedFile.FilePath)) return "";
+
+
+            string readText = File.ReadAllText(sharedFile.FilePath);
+            FileStream fStream = new FileStream(sharedFile.FilePath + encrypt_fileext, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.ReadWrite);
+
+            int length = 0;
+            byte[] toEncrypt = UnicodeEncoding.ASCII.GetBytes(readText);
+            byte[] entropy = CreateRandomEntropy();
+            byte[] encrptedData = ProtectedData.Protect(toEncrypt, entropy, DataProtectionScope.CurrentUser);
+            // Write the encrypted data to a stream.
+            if (fStream.CanWrite && encrptedData != null)
+            {
+                fStream.Write(encrptedData, 0, encrptedData.Length);
+            }
+
+            fStream.Close();
+            return sharedFile.FilePath + encrypt_fileext;
+        }
+
+        private List<CredentialProfile> loadProfileFromCredFile()
+        {
+            List<CredentialProfile> credlist = new List<CredentialProfile>();
+            if (ExistsCredFile())
+            {
+                var crdfile = OpenCredFile();
+                //var sharedFile = new SharedCredentialsFile();
+                if (crdfile != null)
+                {
+                    credlist = crdfile.ListProfiles();// .ListProfileNames();
+                    closeShareProfileFile();    //protect the credential info
+                }
+
+            }
+
+            return credlist;
+        }
+
+        private void closeShareProfileFile()
+        {
+            var sharedFile = new SharedCredentialsFile();
+            string credfilepath = sharedFile.FilePath;
+            if (File.Exists(credfilepath)) File.Delete(credfilepath);   //remove it so no one can open it with notepad, since it's not encrypted
+        }
+
+        private bool ExistsCredFile()
+        {
+            var sharedFile = new SharedCredentialsFile();
+            string encrypt_file = sharedFile.FilePath + encrypt_fileext;
+            if (File.Exists(encrypt_file))
+            {
+                return true;
+            }
+            return false;
+        }
+        private SharedCredentialsFile OpenCredFile()
+        {
+            var sharedFile = new SharedCredentialsFile();
+            string encrypt_file = sharedFile.FilePath + encrypt_fileext;
+            byte[] toDecrypt;
+            if (File.Exists(encrypt_file))
+            {
+                FileStream fStream = new FileStream(sharedFile.FilePath + encrypt_fileext, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+                if (fStream.CanRead)
+                {
+                    toDecrypt = new byte[fStream.Length];
+                    fStream.Read(toDecrypt, 0, (int)fStream.Length);
+
+                    byte[] entropy = CreateRandomEntropy();
+                    byte[] decrypted_bytes = ProtectedData.Unprotect(toDecrypt, entropy, DataProtectionScope.CurrentUser);
+                    string decrypted_readtext = UnicodeEncoding.ASCII.GetString(decrypted_bytes);
+
+                    FileStream fStream1 = new FileStream(sharedFile.FilePath, FileMode.OpenOrCreate);
+
+                    if (fStream1.CanWrite && decrypted_bytes != null)
+                    {
+                        fStream1.Write(decrypted_bytes, 0, decrypted_bytes.Length);
+                        fStream1.Close();
+                    }
+                }
+            }
+
+            return sharedFile;
+        }
+
+        private string unprotectText(string texttodecrypt)
+        {
+            var sharedFile = new SharedCredentialsFile();
+            if (!File.Exists(sharedFile.FilePath + encrypt_fileext)) return "";
+
+            FileStream fStream = new FileStream(sharedFile.FilePath + encrypt_fileext, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+            string unprotectedstring = string.Empty;
+            // Read the encrypted data from a stream.
+            if (fStream.CanRead)
+            {
+                byte[] inBuffer = new byte[fStream.Length];
+                byte[] outBuffer;
+                byte[] entropy = CreateRandomEntropy();
+
+                fStream.Read(inBuffer, 0, Convert.ToInt32(fStream.Length));
+
+                outBuffer = ProtectedData.Unprotect(inBuffer, entropy, DataProtectionScope.CurrentUser);
+                unprotectedstring = UnicodeEncoding.ASCII.GetString(outBuffer);
+            }
+            else
+            {
+                throw new IOException("Could not read the stream.");
+            }
+            return unprotectedstring;
+        }
+
+        public static byte[] CreateRandomEntropy()
+        {
+            // Create a byte array to hold the random value.
+            //byte[] entropy = new byte[16];
+
+            // Create a new instance of the RNGCryptoServiceProvider.
+            // Fill the array with a random value.
+            //new RNGCryptoServiceProvider().GetBytes(entropy);
+            byte[] s_additionalEntropy = { 9, 8, 7, 6, 5 };
+            // Return the array.
+            return s_additionalEntropy;
+        }
+
+        private string createEmptyCredentialFile()
+        {
+            CredentialProfileOptions options = new CredentialProfileOptions
+            {
+                AccessKey = ""
+            };
+            string dumstr = "dummy";
+            var profile = new CredentialProfile(dumstr, options);
+            var sharedFile = new SharedCredentialsFile();
+            sharedFile.RegisterProfile(profile);
+            //sharedFile.UnregisterProfile(dumstr);
+            string filepath = sharedFile.FilePath;
+            return filepath;
+        }
+        private void WriteProfile(string profileName, string keyId, string secret, string acctExternalId, string dbadmusr, string dbadmpwd)
+        {
+            var sharedFile = new SharedCredentialsFile();
+            CredentialProfileOptions options = new CredentialProfileOptions
+            {
+                AccessKey = keyId,
+                SecretKey = secret
+            };
+            var profile = new CredentialProfile(profileName, options);
+
+            if (ExistsCredFile())
+            {
+                sharedFile = OpenCredFile();
+                sharedFile.RegisterProfile(profile);
+            }
+            else
+            {
+                sharedFile.RegisterProfile(profile);
+            }
+
+            saveCredFile();
+            closeShareProfileFile();    //protect the credential info
+
+        }
+
+        private void btnSaveCred_Click(object sender, EventArgs e)
+        {
+            if (btnSaveCred.Text.ToUpper().Contains("USE PROFILE"))
+            {
+                string oodbendpoint = string.Empty;
+                oodbendpoint = AWSConnections.getAWSRDSCleint(tbxAccKey.Text, tbxSecKey.Text, tbxRDSsf.Text);
+                tbxDBep.Text = oodbendpoint;
+
+
+                string oos3name = string.Empty;
+                oos3name = AWSConnections.getAWSS3Client(tbxAccKey.Text, tbxSecKey.Text);
+                tbxS3buckets.Text = oos3name;
+
+                string oousernmae = string.Empty;
+                oousernmae = AWSConnections.LoginDB(tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkDBName);
+
+            }
+            else if (btnSaveCred.Text.ToUpper().Contains("CREATE PROFILE"))
+            {
+
+                string oodbendpoint = string.Empty;
+                oodbendpoint = AWSConnections.getAWSRDSCleint(tbxAccKey.Text, tbxSecKey.Text, tbxRDSsf.Text);
+                tbxDBep.Text = oodbendpoint;
+
+                string oos3name = string.Empty;
+                oos3name = AWSConnections.getAWSS3Client(tbxAccKey.Text, tbxSecKey.Text);
+                tbxS3buckets.Text = oos3name;
+
+                string oousernmae = string.Empty;
+                oousernmae = AWSConnections.LoginDB(tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkDBName);
+
+                if (oousernmae != tbxDbUserP.Text)
+                {
+                    ShowMsgD("Database Login Failed. Please try again.");
+                    return;  //login failed
+                }
+                //bool b_dbtbl_create = AWSConnections.CreateooTables(ooData.olinkDBName, tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text);
+                //if (!AWSConnections.checkooDbExists(ooData.olinkDBName, tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkRDSDBName))
+                //{
+                //    bool b_dbtb//l_create = AWSConnections.CreateooDb(ooData.olinkDBName, tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkRDSDBName);
+                //}
+
+                WriteProfile(tbxProfName.Text, tbxAccKey.Text, tbxSecKey.Text, "", tbxDbUserP.Text, tbxDbPwdP.Text);
+                listprofilenames();
+            }
+            else if (btnSaveCred.Text.ToUpper().Contains("LOG IN"))
+            {
+                string oodbendpoint = string.Empty;
+                oodbendpoint = AWSConnections.getAWSRDSCleint(tbxAccKey.Text, tbxSecKey.Text, tbxRDSsf.Text);
+                tbxDBep.Text = oodbendpoint;
+
+                string oousernmae = string.Empty;
+                oousernmae = AWSConnections.LoginDB(tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkDBName);
+                if (oousernmae != tbxDbUserP.Text)
+                {
+                    ShowMsgD("Database server Login Failed. Please try again.");
+                    return;  //login failed
+                }
+                ShowMsgD("Remote database server login done.");
+
+                if (!AWSConnections.checkooDbExists(ooData.olinkDBName, tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkRDSDBName))
+                {
+                    ShowMsgD("No oLink database has been created to proceed. Please try again.");
+                    return;
+                }
+                ShowMsgD("Default database to function has been found.");
+
+                if (!AWSConnections.ExistsooTables(ooData.olinkDBName, tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text))
+                {
+                    ShowMsgD("Not all default oLink database tables have been created to proceed. Please try again.");
+                    return;
+                }
+                ShowMsgD("Default database tables to function have been found.");
+
+                string conn = AWSConnections.RDSConnString(tbxDbUserP.Text, tbxDbPwdP.Text, tbxDBep.Text, ooData.olinkDBName);
+                tbxDbConnString.Text = conn;
+
+                ShowhideDataTabs(true);
+                ShowhideCredsTab(false);
+                ShowDataGridView1();
+            }
+
+        }
+
+        private CredentialProfile retrieveByProfileName(string profilename)
+        {
+            List<CredentialProfile> pl = loadProfileFromCredFile();
+            var profile = pl.Find(p => p.Name == profilename);
+            return profile;
+        }
+        private void cbxProfiles_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            if (cbxProfiles.SelectedIndex > -1)
+            {
+                CredentialProfile profile = retrieveByProfileName(cbxProfiles.Text);
+
+                if (profile != null && profile.Options != null)
+                {
+                    tbxAccKey.Text = profile.Options.AccessKey;
+                    tbxSecKey.Text = profile.Options.SecretKey;
+                    EnableDbLogins(true);
+                    btnSaveCred.Enabled = true;
+                    btnSaveCred.Text = "Log In";
+                    tbxProfName.Text = cbxProfiles.Text;
+                    tbxcurUser.Text = AWSConnections.getAWSUserId(tbxAccKey.Text, tbxSecKey.Text);
+                    tbxS3buckets.Text = AWSConnections.getAWSS3Client(tbxAccKey.Text, tbxSecKey.Text);
+                    tbxDBep.Text = AWSConnections.getAWSRDSCleint(tbxAccKey.Text, tbxSecKey.Text, tbxRDSsf.Text);
+                }
+            }
+        }
+
+        private void tabEntries_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            if (tabEntries.SelectedIndex >= 0)
+            {//Management Console
+                textBoxS3Id.Text = tbxAccKey.Text;
+                textBoxS3Key.Text = tbxSecKey.Text;
+                textBoxS3Bucket.Text = tbxS3buckets.Text;
+            }
+        }
+
+        private void ckbcloudfront_CheckedChanged(object sender, EventArgs e)
+        {
+            if (!ckbcloudfront.Checked)
+            {
+                DialogResult confirmResult = MessageBox.Show("Uncheck Cloudfront will make your S3 bucket public",
+                                     "Confirm Uncheck!!",
+                                     MessageBoxButtons.YesNo);
+                if (confirmResult == DialogResult.No)
+                {
+                    ckbcloudfront.Checked = true;
+                }
+            }
+        }
+
+        private void TogglePublicAccessBlock(bool toblock)
+        {
+            try
+            {
+                //When using CF: Turn on Block All Public Access" & Deny all others from GetS3Object (Permission Policy)
+                //When ot using CF: Turn off "Block All Public Access" & Allow all others from GetS3Object (Permission Policy)
+                // Create a client
+                AmazonS3Client s3Client = AWSConnections.getAWSS3Cli(tbxAccKey.Text, tbxSecKey.Text);
+
+                string st_myS3name = AWSConnections.getAWSS3Client(tbxAccKey.Text, tbxSecKey.Text);
+
+                GetPublicAccessBlockRequest request = new GetPublicAccessBlockRequest();
+                request.BucketName = st_myS3name;
+                GetPublicAccessBlockResponse response = s3Client.GetPublicAccessBlock(request);
+
+                PutPublicAccessBlockRequest request1 = new PutPublicAccessBlockRequest();
+                request1.BucketName = st_myS3name;
+                request1.PublicAccessBlockConfiguration = new PublicAccessBlockConfiguration
+                { BlockPublicAcls = toblock, BlockPublicPolicy = toblock, IgnorePublicAcls = toblock, RestrictPublicBuckets = toblock };
+                PutPublicAccessBlockResponse response1 = s3Client.PutPublicAccessBlock(request1);
+                string msg_x = toblock ? "private." : "public.";
+                ShowMsgD("S3 Bucket is made " + msg_x);
+                s3Client.Dispose();
+
+            }
+            catch (Exception exp)
+            {
+                ShowMsgD("S3 Bucketsetting change failed. Please manually modify on your account if necessary. ");
+            }
+
         }
     }
 }
