+++ oLink/packages.config
@@ -1,11 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="AngleSharp" version="0.17.1" targetFramework="net48" />
-  <package id="AngleSharp.Css" version="0.16.4" targetFramework="net48" />
-  <package id="HtmlSanitizer" version="7.1.542" targetFramework="net48" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net48" />
-  <package id="System.Text.Encoding.CodePages" version="6.0.0" targetFramework="net48" />
+  <package id="AngleSharp" version="0.17.1" targetFramework="net472" />
+  <package id="AngleSharp.Css" version="0.16.4" targetFramework="net472" />
+  <package id="AWSSDK.CloudFront" version="3.7.8.2" targetFramework="net48" />
+  <package id="AWSSDK.Core" version="3.7.12.26" targetFramework="net48" />
+  <package id="AWSSDK.IdentityManagement" version="3.7.2.166" targetFramework="net48" />
+  <package id="AWSSDK.RDS" version="3.7.15.3" targetFramework="net48" />
+  <package id="AWSSDK.S3" version="3.7.9.33" targetFramework="net48" />
+  <package id="HtmlSanitizer" version="7.1.542" targetFramework="net472" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" />
+  <package id="System.Text.Encoding.CodePages" version="6.0.0" targetFramework="net472" />
 </packages>
